% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{abbreviations}
\alias{abbreviations}
\title{Abbreviations used in the \strong{fokus} package}
\usage{
abbreviations(expand = FALSE)
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} listing an opinionated set of abbreviations used in the \R code and documentation of the \strong{fokus} package.
}
