% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{abbrs}
\alias{abbrs}
\title{Abbreviations used in the \strong{fokus} package}
\usage{
abbrs(unnest = FALSE)
}
\arguments{
\item{unnest}{Whether to unnest the \code{full_expression} column and return the data in long format. If \code{FALSE}, a "nested" list column \code{full_expressions} will
be returned, meaning the values in column \code{abbreviation} will be unique.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a \link[tibble:tbl_df-class]{tibble} listing an opinionated set of abbreviations used in the \R code and documentation of the \strong{fokus} package.
}
\keyword{internal}
