% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{add_vars_to_combos}
\alias{add_vars_to_combos}
\title{Add variable names to combination list}
\usage{
add_vars_to_combos(list, var_names)
}
\arguments{
\item{list}{A list as returned by the \verb{combos_*()} functions like \code{\link[=combos_ballot_types]{combos_ballot_types()}}.}

\item{var_names}{A character vector of variable names.}
}
\value{
A list with \code{list} × \code{var_names} elements.
}
\description{
Adds one or more \code{var_name}s to a combination list.
}
\examples{
fokus::combos_proposals(ballot_date = "2023-06-18",
                        canton = "aargau",
                        incl_nr = FALSE) |>
  fokus::add_vars_to_combos(var_names = c("age_group", "favored_party"))
}
\seealso{
Other combination functions:
\code{\link{combos_ballot_types}()},
\code{\link{combos_elections}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()}

Other variable-related functions:
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_lvls}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_title}()},
\code{\link{var_val_set}()}
}
\concept{combo}
\concept{vars}
