% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_argument_sides}
\alias{all_argument_sides}
\title{Referendum proposal argument sides}
\format{
A character vector.
}
\usage{
all_argument_sides
}
\description{
A vector of all possible referendum proposal argument sides.
}
\examples{
fokus::all_argument_sides
}
\seealso{
Other FOKUS metadata functions:
\code{\link{all_ballot_dates}},
\code{\link{all_ballot_types}},
\code{\link{all_cantons}},
\code{\link{all_lvls}},
\code{\link{all_postal_dispatch_types}},
\code{\link{all_postal_dispatch_ways}},
\code{\link{all_prcds}},
\code{\link{all_proposal_types}},
\code{\link{all_response_option_types}}
}
\concept{metadata}
\keyword{datasets}
