% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_ballot_dates}
\alias{all_ballot_dates}
\title{FOKUS-covered ballot dates}
\format{
A vector of \link[base:Dates]{dates}.
}
\usage{
all_ballot_dates
}
\description{
A vector of ballot dates covered by FOKUS surveys up until 2021-11-28.
}
\examples{
fokus::all_ballot_dates
}
\seealso{
\code{\link{all_cantons}} \code{\link{all_ballot_types}} \code{\link{all_lvls}} \code{\link{all_prcds}}
\code{\link{all_proposal_types}} \code{\link{all_response_option_types}}
\code{\link{all_argument_sides}}
}
\keyword{datasets}
