% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_cantons}
\alias{all_cantons}
\title{FOKUS-covered cantons}
\format{
A character vector.
}
\usage{
all_cantons
}
\description{
A vector of all \link{cantons} covered by FOKUS surveys.
}
\examples{
fokus::all_cantons
}
\seealso{
\code{\link{all_ballot_dates}} \code{\link{all_ballot_types}} \code{\link{all_lvls}} \code{\link{all_prcds}}
\code{\link{all_proposal_types}} \code{\link{all_response_option_types}} \code{\link{all_argument_sides}}
}
\keyword{datasets}
