% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_lvls}
\alias{all_lvls}
\title{Political levels}
\format{
A character vector.
}
\usage{
all_lvls
}
\description{
A vector of all possible \link[=lvls]{political levels}.
}
\examples{
fokus::all_lvls
}
\seealso{
\code{\link{all_ballot_dates}} \code{\link{all_cantons}} \code{\link{all_ballot_types}} \code{\link{all_prcds}}
\code{\link{all_proposal_types}} \code{\link{all_response_option_types}} \code{\link{all_argument_sides}}
}
\keyword{datasets}
