% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_postal_dispatch_ways}
\alias{all_postal_dispatch_ways}
\title{Postal dispatch ways}
\format{
A character vector.
}
\usage{
all_postal_dispatch_ways
}
\description{
A vector of all possible \link[=postal_dispatch_way]{postal dispatch ways}.
}
\examples{
fokus::all_postal_dispatch_ways
}
\seealso{
Other FOKUS metadata functions:
\code{\link{all_argument_sides}},
\code{\link{all_ballot_dates}},
\code{\link{all_ballot_types}},
\code{\link{all_cantons}},
\code{\link{all_lvls}},
\code{\link{all_postal_dispatch_types}},
\code{\link{all_prcds}},
\code{\link{all_proposal_types}},
\code{\link{all_response_option_types}}
}
\concept{metadata}
\keyword{datasets}
