% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_proposal_types}
\alias{all_proposal_types}
\title{Referendum proposal types}
\format{
A character vector.
}
\usage{
all_proposal_types
}
\description{
A vector of all possible \link[=proposal_type]{referendum proposal types}.
}
\examples{
fokus::all_proposal_types
}
\seealso{
Other FOKUS metadata functions:
\code{\link{all_argument_sides}},
\code{\link{all_ballot_dates}},
\code{\link{all_ballot_types}},
\code{\link{all_cantons}},
\code{\link{all_lvls}},
\code{\link{all_postal_dispatch_types}},
\code{\link{all_postal_dispatch_ways}},
\code{\link{all_prcds}},
\code{\link{all_response_option_types}},
\code{\link{q_item_keys}}
}
\concept{metadata}
\keyword{datasets}
