% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_qstn_groups}
\alias{all_qstn_groups}
\title{Standardized referendum proposal question groups}
\format{
A character vector.
}
\usage{
all_qstn_groups
}
\description{
A vector of all possible referendum \link[=proposal_qstn_groups]{proposal question groups}.
}
\examples{
fokus::all_qstn_groups
}
\seealso{
Other FOKUS metadata:
\code{\link{all_argument_sides}},
\code{\link{all_ballot_dates}},
\code{\link{all_ballot_types}},
\code{\link{all_cantons}},
\code{\link{all_election_seat_types}},
\code{\link{all_langs}},
\code{\link{all_lvls}},
\code{\link{all_main_motive_types}},
\code{\link{all_name_types}},
\code{\link{all_postal_dispatch_types}},
\code{\link{all_postal_dispatch_ways}},
\code{\link{all_prcds}},
\code{\link{all_proposal_types}},
\code{\link{all_response_option_types}},
\code{\link{qstnr_item_keys}}
}
\concept{metadata}
\keyword{datasets}
