% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{all_response_option_types}
\alias{all_response_option_types}
\title{Response option types}
\format{
A character vector.
}
\usage{
all_response_option_types
}
\description{
A vector of all possible response option types defined in the \link[=raw_q]{raw FOKUS questionnaire data}.
}
\examples{
fokus::all_response_option_types
}
\seealso{
\code{\link{all_ballot_dates}} \code{\link{all_cantons}} \code{\link{all_ballot_types}} \code{\link{all_lvls}}
\code{\link{all_prcds}} \code{\link{all_proposal_types}}
}
\keyword{datasets}
