% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{argument_side_prose}
\alias{argument_side_prose}
\title{Transform argument side into prose string}
\usage{
argument_side_prose(side, lang = all_langs)
}
\arguments{
\item{side}{Proposal argument side. One of \code{"pro"} or \code{"contra"}.}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=all_argument_sides]{proposal argument} side into a full prose string representation.
}
\examples{
fokus::all_argument_sides |>
  purrr::map(\(x) tibble::tibble(side = x,
                                 de = fokus::argument_side_prose(side = x,
                                                                 lang = "de"),
                                 en = fokus::argument_side_prose(side = x,
                                                                 lang = "en"))) |>
  purrr::list_rbind()
}
\seealso{
Other prose text functions:
\code{\link{ballot_type_prose}()},
\code{\link{lvl_prose}()},
\code{\link{motive_type_prose}()},
\code{\link{postal_dispatch_type_prose}()},
\code{\link{postal_dispatch_way_prose}()},
\code{\link{proposal_name_declined_de}()}
}
\concept{prose}
