% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{auth_g_drive}
\alias{auth_g_drive}
\title{Authorize googledrive}
\usage{
auth_g_drive(
  path_gcp_service_account_key = path_private("config/gcp_service_account_key.json")
)
}
\arguments{
\item{path_gcp_service_account_key}{Path to the GCP Service Account Key JSON file.}
}
\value{
\code{path_gcp_service_account_key}, invisibly.
}
\description{
Authorizes the googledrive package to access and manage files on your Google Drive via a \href{https://cloud.google.com/iam/docs/creating-managing-service-account-keys}{Google Cloud Platform (GCP) Service Account Key} file in JSON format.
}
\seealso{
Other Google Apps functions
\code{\link{upload_to_g_drive}()}
}
\concept{g_apps}
\keyword{internal}
