% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{auth_g_sheets_gcp}
\alias{auth_g_sheets_gcp}
\title{Authorize googlesheets4 using GCP Service Account Key}
\usage{
auth_g_sheets_gcp(
  path_gcp_service_account_key = Sys.getenv("PATH_GCP_KEY_ZDA")
)
}
\arguments{
\item{path_gcp_service_account_key}{Path to the GCP Service Account Key JSON file.}
}
\value{
\code{path_gcp_service_account_key}, invisibly.
}
\description{
Authorizes the googlesheets4 package to access and manage Google Sheets via a \href{https://cloud.google.com/iam/docs/creating-managing-service-account-keys}{Google Cloud Platform (GCP) Service Account Key} file (in JSON format). See the \href{https://gargle.r-lib.org/articles/non-interactive-auth.html#provide-a-service-account-token-directly}{relevant documentation} for details.
}
\details{
The recommended way to provide the filesystem path to the GCP Service Account Key file is to set the \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variable} \code{PATH_GCP_KEY_ZDA} to that path, e.g. via the \href{https://rstats.wtf/r-startup.html#renviron}{\code{.Renviron} file}.
}
\seealso{
Other Google Apps functions:
\code{\link{auth_g_drive_gcp}()},
\code{\link{backup_g_file}()},
\code{\link{backup_g_sheet}()},
\code{\link{g_file_mod_time}()},
\code{\link{upload_to_g_drive}()}
}
\concept{g_apps}
\keyword{internal}
