% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{backup_g_file}
\alias{backup_g_file}
\title{Backup Google Drive file locally}
\usage{
backup_g_file(
  g_id,
  path,
  type = NULL,
  path_gcp_service_account_key = Sys.getenv("PATH_GCP_KEY_ZDA"),
  overwrite = TRUE,
  force = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{g_id}{Google Drive file ID. A character scalar.}

\item{path}{Path to the local file backup destination. A character scalar.}

\item{type}{Character. Only consulted if \code{file} is a native Google file.
Specifies the desired type of the exported file. Will be processed via
\code{\link[googledrive:drive_mime_type]{drive_mime_type()}}, so either a file extension like \code{"pdf"} or a full MIME
type like \code{"application/pdf"} is acceptable.}

\item{path_gcp_service_account_key}{Path to the GCP Service Account Key JSON file. See \code{\link[=auth_g_drive_gcp]{auth_g_drive_gcp()}} for details.}

\item{overwrite}{Whether or not to overwrite an already existing file under \code{path}.}

\item{force}{Whether or not to force overwriting the file regardless whether it has changed since the last backup or not.}

\item{quiet}{Whether or not to \link[googledrive:googledrive-configuration]{suppress printing status output from googledrive operations}.}
}
\value{
An object of class \link[googledrive:dribble]{dribble}, a tibble with one row per file if local backup was (over)written, otherwise \code{NULL}, meaning
remote file hasn't deviated from local backup since last run, invisibly.
}
\description{
Creates a local backup of a \href{https://en.wikipedia.org/wiki/Google_Drive}{Google Drive} file.
}
\details{
Essentially a convenience wrapper around \code{\link[googledrive:drive_download]{googledrive::drive_download()}}.

If a \href{https://en.wikipedia.org/wiki/Google_Sheets}{Google Sheet} that includes multiple worksheets is backed up to a file \code{type} that doesn't support
worksheets like \code{csv}, only the default (first) worksheet is written. In contrast, \code{\link[=backup_g_sheet]{backup_g_sheet()}} allows to specify any worksheet to be backed up (but
always writes a single worksheet only regardless of the filetype).
}
\seealso{
Other Google Apps functions:
\code{\link{auth_g_drive_gcp}()},
\code{\link{auth_g_sheets_gcp}()},
\code{\link{backup_g_sheet}()},
\code{\link{g_file_mod_time}()},
\code{\link{upload_to_g_drive}()}
}
\concept{g_apps}
