% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{backup_g_sheet}
\alias{backup_g_sheet}
\title{Backup Google Sheet locally}
\usage{
backup_g_sheet(
  g_id,
  path,
  ...,
  path_gcp_service_account_key = Sys.getenv("PATH_GCP_KEY_ZDA"),
  overwrite = TRUE,
  force = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{g_id}{Google Drive file ID. A character scalar.}

\item{path}{Path to the local file backup destination. A character scalar.}

\item{...}{Further arguments passed on to \code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}

\item{path_gcp_service_account_key}{Path to the GCP Service Account Key JSON file. See \code{\link[=auth_g_drive_gcp]{auth_g_drive_gcp()}} for details.}

\item{overwrite}{Whether or not to overwrite an already existing file under \code{path}.}

\item{force}{Whether or not to force overwriting the file regardless whether it has changed since the last backup or not.}

\item{quiet}{Whether or not to \link[googledrive:googledrive-configuration]{suppress printing status output from googledrive} and \link[googlesheets4:googlesheets4-configuration]{googlesheets4 operations}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} if the local backup was (over)written, otherwise \code{NULL}, meaning the remote file hasn't deviated from the local backup
since the last run, invisibly.
}
\description{
Creates a local backup of a \href{https://en.wikipedia.org/wiki/Google_Sheets}{Google Sheet}. It is accessed via \code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}} and written via
\code{\link[readr:write_delim]{readr::write_csv()}} or \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}, depending on the file extension of \code{path}.
}
\details{
\code{backup_g_sheet()} only backs up a single worksheet at once (specified by the optional \code{sheet} argument). If you intend to backup multiple worksheets of the
same Google Sheet, consider using \code{\link[=backup_g_file]{backup_g_file()}} in combination with a file \code{type} that supports multiple worksheets like \code{"ods"} or \code{"xlsx"}.
}
\seealso{
Other Google Apps functions:
\code{\link{auth_g_drive_gcp}()},
\code{\link{auth_g_sheets_gcp}()},
\code{\link{backup_g_file}()},
\code{\link{g_file_mod_time}()},
\code{\link{upload_to_g_drive}()}
}
\concept{g_apps}
