% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{ballot_dates}
\alias{ballot_dates}
\title{List FOKUS-covered ballot dates}
\format{
A vector of \link[base:Dates]{dates}.
}
\usage{
ballot_dates
}
\description{
A vector of ballot dates covered by FOKUS surveys up until 2021-11-28.
}
\seealso{
\code{\link{cantons}} \code{\link{ballot_metadata}} \code{\link{response_option_types}} \code{\link{q_item_keys}}
\code{\link{proposal_types}}
}
\keyword{datasets}
