% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{ballot_metadata}
\alias{ballot_metadata}
\title{FOKUS ballot-date-canton metadata}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
ballot_metadata
}
\description{
A tibble of FOKUS-survey-related ballot-date-canton metadata, valid up until 2021-11-28.
}
\seealso{
\code{\link{ballot_dates}} \code{\link{cantons}} \code{\link{response_option_types}} \code{\link{q_item_keys}}
\code{\link{proposal_types}}
}
\keyword{datasets}
