% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{ballot_title}
\alias{ballot_title}
\title{Get ballot title}
\usage{
ballot_title(ballot_date, canton, lang = c("de", "en"))
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character scalar.
}
\description{
Returns the ballot title consisting of the \link[=ballot_types]{ballot type} and the ballot date in German prose.
}
\examples{
ballot_title(ballot_date = "2019-10-20",
             canton = "aargau")
}
