% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{ballot_type_prose}
\alias{ballot_type_prose}
\title{Transform ballot type into prose string}
\usage{
ballot_type_prose(ballot_type = all_ballot_types, lang = c("de", "en"))
}
\arguments{
\item{ballot_type}{Ballot type. One of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=all_ballot_types]{ballot type} into a full prose string representation.
}
\examples{
fokus::all_ballot_types |>
  purrr::map(\(x) tibble::tibble(type = x,
                                 de = fokus::ballot_type_prose(ballot_type = x,
                                                               lang = "de"),
                                 en = fokus::ballot_type_prose(ballot_type = x,
                                                               lang = "en"))) |>
  purrr::list_rbind()
}
\seealso{
Other prose text functions:
\code{\link{postal_dispatch_type_prose}()},
\code{\link{postal_dispatch_way_prose}()}
}
\concept{prose}
