% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{ballot_types}
\alias{ballot_types}
\title{Determine ballot types}
\usage{
ballot_types(ballot_date = all_ballot_dates, canton = all_cantons)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A character vector of ballot types. One or more of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}
}
\description{
Determines the types of the ballot covered by the FOKUS survey for the specified canton at the specified ballot date.
}
\examples{
fokus::ballot_types(ballot_date = "2018-09-23",
                    canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{cantons}()},
\code{\link{has_ballot_types}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()}
}
\concept{predicate_fundamental}
