% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{ballots}
\alias{ballots}
\title{Ballot data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
ballots
}
\description{
A list of tibbles containing official statistical data about all FOKUS-covered ballots. The data includes i.a. the total electorate as well as participation
numbers per voting channel and per voting decision.
}
\examples{
fokus::ballots
}
\seealso{
Other datasets:
\code{\link{elections}},
\code{\link{proposals}},
\code{\link{qstnrs}}
}
\concept{qstnr_data}
\keyword{datasets}
