% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\docType{data}
\name{cantons}
\alias{cantons}
\title{List cantons that are covered in \emph{any} FOKUS survey}
\format{
A character vector.
}
\usage{
cantons
}
\description{
Simply returns a vector of cantons that were part of at least one FOKUS survey up until 2020-10-23.
}
\seealso{
\link{ballot_dates} \link{ballot_metadata}
}
\keyword{datasets}
