% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_ballot_types}
\alias{combos_ballot_types}
\title{Get ballot type combinations}
\usage{
combos_ballot_types(
  ballot_dates = all_ballot_dates,
  lvls = all_lvls,
  cantons = all_cantons,
  ballot_types = all_ballot_types,
  incl_lvl = TRUE
)
}
\arguments{
\item{ballot_dates}{FOKUS-covered ballot date(s). One or more of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{cantons}{FOKUS-covered canton name(s). One or more of
\itemize{
\item \code{"aargau"}
}}

\item{ballot_types}{Ballot type(s). One or more of \code{"referendum"} or \code{"election"}}

\item{incl_lvl}{Whether or not to include the political levels in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations.}
}
\value{
A list with an element per ballot-type and optionally -political-level combination.
}
\description{
Returns a list with a metadata item per ballot type, ballot date, canton and optionally political level.
}
\examples{
# of all covered ballots
fokus::combos_ballot_types()

# only of 2023-06-18 ballot in aargau
fokus::combos_ballot_types(ballot_dates = "2023-06-18",
                           cantons = "aargau")
}
\seealso{
Other combination functions:
\code{\link{add_vars_to_combos}()},
\code{\link{combos_elections}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()}

Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{is_representative}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()},
\code{\link{survey_channels}()}
}
\concept{combo}
\concept{predicate_fundamental}
