% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_elections}
\alias{combos_elections}
\title{Get election combinations}
\usage{
combos_elections(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  incl_prcd = TRUE,
  incl_nr = incl_prcd
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{incl_prcd}{Whether or not to include election procedures in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations.}

\item{incl_nr}{Whether or not to include election numbers in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations. Only
relevant if \code{incl_prcd = TRUE}.}
}
\value{
A list with an element per political-level and optionally -election-procedure and -election-number combination.
}
\description{
Returns a list containing the political levels, election procedures and election numbers for all elections that have been covered by the FOKUS survey for the
specified canton at the specified ballot date on the specified political level(s).
}
\examples{
fokus::combos_elections(ballot_date = "2019-10-20",
                        canton = "aargau")

# without election numbers
fokus::combos_elections(ballot_date = "2019-10-20",
                        canton = "aargau",
                        incl_prcd = TRUE,
                        incl_nr = FALSE)

# without election procedures and numbers
fokus::combos_elections(ballot_date = "2019-10-20",
                        canton = "aargau",
                        incl_prcd = FALSE,
                        incl_nr = FALSE)
}
\seealso{
Other election predicate functions:
\code{\link{election_candidates}()},
\code{\link{election_name}()},
\code{\link{election_names_combined}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
