% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_proposal_arguments}
\alias{combos_proposal_arguments}
\title{Get referendum proposal combinations for which arguments have been queried}
\usage{
combos_proposal_arguments(
  ballot_dates = all_ballot_dates,
  lvls = all_lvls,
  cantons = all_cantons,
  incl_side = TRUE,
  incl_argument_nr = incl_side
)
}
\arguments{
\item{ballot_dates}{FOKUS-covered ballot date(s). One or more of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{cantons}{FOKUS-covered canton name(s). One or more of
\itemize{
\item \code{"aargau"}
}}

\item{incl_side}{Whether or not to include argument sides (pro/contra) in the resulting list. Setting this to \code{FALSE} potentially results in fewer
combinations.}

\item{incl_argument_nr}{Whether or not to include argument numbers in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations.
Setting this to \code{TRUE} implies \code{incl_side = TRUE}.}
}
\value{
A list with an element per political-level-proposal-number and optionally -argument-side and -argument-number combination.
}
\description{
Returns a list with a metadata item per ballot date, political level, canton, proposal number and optionally proposal argument side (pro/contra) and number.
}
\examples{
fokus::combos_proposal_arguments(ballot_dates = "2023-06-18",
                                 cantons = "aargau")
# without argument numbers
fokus::combos_proposal_arguments(ballot_dates = "2023-06-18",
                                 cantons = "aargau",
                                 incl_argument_nr = FALSE)
# without argument sides and numbers
fokus::combos_proposal_arguments(ballot_dates = "2023-06-18",
                                 cantons = "aargau",
                                 incl_side = FALSE,
                                 incl_argument_nr = FALSE)
}
\seealso{
Other combination functions:
\code{\link{add_vars_to_combos}()},
\code{\link{combos_ballot_types}()},
\code{\link{combos_elections}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()}

Other proposal predicate functions:
\code{\link{argument_proposal_nrs}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()},
\code{\link{has_proposal_arguments}()},
\code{\link{has_proposal_main_motives}()},
\code{\link{is_proposal_name_plural}()},
\code{\link{main_motive_proposal_nrs}()},
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_type}()}
}
\concept{combo}
\concept{predicate_proposal}
