% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_proposal_main_motives}
\alias{combos_proposal_main_motives}
\title{Get referendum proposal combinations for which main motives have been queried}
\usage{
combos_proposal_main_motives(
  ballot_dates = all_ballot_dates,
  lvls = all_lvls,
  cantons = all_cantons,
  incl_type = TRUE
)
}
\arguments{
\item{ballot_dates}{FOKUS-covered ballot date(s). One or more of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{cantons}{FOKUS-covered canton name(s). One or more of
\itemize{
\item \code{"aargau"}
}}

\item{incl_type}{Whether or not to include motive types (yes/no) in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations.}
}
\value{
A list with an element per political-level-proposal-number and optionally -type combination.
}
\description{
Returns a list with a metadata item per ballot date, political level, canton, proposal number and optionally motive type for which \link[=proposal_main_motives]{main motives} have been queried in the respective post-voting survey.
}
\examples{
fokus::combos_proposal_main_motives(ballot_dates = "2023-06-18",
                                    cantons = "aargau")
# without types
fokus::combos_proposal_main_motives(ballot_dates = "2023-06-18",
                                    cantons = "aargau",
                                    incl_type = FALSE)
}
\seealso{
Other combination functions:
\code{\link{add_vars_to_combos}()},
\code{\link{combos_ballot_types}()},
\code{\link{combos_elections}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposals}()}

Other proposal predicate functions:
\code{\link{argument_proposal_nrs}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposals}()},
\code{\link{has_proposal_arguments}()},
\code{\link{has_proposal_main_motives}()},
\code{\link{is_proposal_name_plural}()},
\code{\link{main_motive_proposal_nrs}()},
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_type}()}
}
\concept{combo}
\concept{predicate_proposal}
