% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_proposals}
\alias{combos_proposals}
\title{Get referendum proposal combinations}
\usage{
combos_proposals(
  ballot_dates = all_ballot_dates,
  lvls = all_lvls,
  cantons = all_cantons,
  incl_nr = TRUE
)
}
\arguments{
\item{ballot_dates}{FOKUS-covered ballot date(s). One or more of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{cantons}{FOKUS-covered canton name(s). One or more of
\itemize{
\item \code{"aargau"}
}}

\item{incl_nr}{Whether or not to include proposal numbers in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations.}
}
\value{
A list with an element per political-level and optionally -proposal-number combination.
}
\description{
Returns a list with a metadata item per ballot date, political level, canton and optionally proposal number.
}
\examples{
fokus::combos_proposals(ballot_dates = "2023-06-18",
                        cantons = "aargau")

# without proposal numbers
fokus::combos_proposals(ballot_dates = "2023-06-18",
                        cantons = "aargau",
                        incl_nr = FALSE)
}
\seealso{
Other combination functions:
\code{\link{add_vars_to_combos}()},
\code{\link{combos_ballot_types}()},
\code{\link{combos_elections}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()}

Other proposal predicate functions:
\code{\link{argument_proposal_nrs}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{has_proposal_arguments}()},
\code{\link{has_proposal_main_motives}()},
\code{\link{is_proposal_name_plural}()},
\code{\link{main_motive_proposal_nrs}()},
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_type}()}
}
\concept{combo}
\concept{predicate_proposal}
