% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_candidate_string}
\alias{election_candidate_string}
\title{Assemble majoritarian election's candidate string(s)}
\usage{
election_candidate_string(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  election_nr = 1L,
  candidate_nrs = NULL,
  incl_party = TRUE
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{Election number. A positive integer scalar (in almost all cases \code{1L}).}

\item{candidate_nrs}{Election candidate numbers to include. A vector of positive integers or \code{NULL}. If \code{NULL}, all candidates will be included.}

\item{incl_party}{Whether or not to include the candidate's political party in the resulting string (in parentheses).}
}
\value{
A character vector.
}
\description{
Assembles one or more majoritarian election candidate string(s) consisting of the candidate's first name, last name and optionally political party (in
parentheses).
}
\examples{
fokus::election_candidate_string(ballot_date = "2019-10-20",
                                 lvl = "cantonal",
                                 canton = "aargau",
                                 candidate_nrs = 1:3)
}
\seealso{
Other election predicate functions
\code{\link{election_candidates}()},
\code{\link{election_names_combined}()},
\code{\link{election_name}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
