% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_candidates}
\alias{election_candidates}
\title{Get election candidates}
\usage{
election_candidates(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  election_nr = 1L
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{The election number. A positive integer scalar (in almost all cases \code{1L}).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns the name and party of all candidates running for the specified majoritarian election.
}
\examples{
fokus::election_candidates(ballot_date = "2019-10-20",
                           lvl = "cantonal",
                           canton = "aargau",
                           election_nr = 1)
}
