% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_parties}
\alias{election_parties}
\title{Get proportional election's political parties}
\usage{
election_parties(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  election_nr = 1L,
  past = FALSE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{Election number. A positive integer scalar (in almost all cases \code{1L}).}

\item{past}{Whether to extract the current or the predecessor election's parties.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns the questionnaire code as well as different versions of the name of all parties for the specified proportional election.
}
\examples{
fokus::election_parties(ballot_date = "2019-10-20",
                        lvl = "federal",
                        canton = "aargau")

fokus::election_parties(ballot_date = "2019-10-20",
                        lvl = "federal",
                        canton = "aargau",
                        past = TRUE)
}
\seealso{
Other election predicate functions:
\code{\link{election_candidate_string}()},
\code{\link{election_candidates}()},
\code{\link{election_names_combined}()},
\code{\link{election_name}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
