% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_tickets}
\alias{election_tickets}
\title{Get proportional election's tickets}
\usage{
election_tickets(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons(ballot_date),
  election_nr = 1L
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{election_nr}{Election number. A positive integer scalar (in almost all cases \code{1L}).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns the name, number and party of all tickets for the specified proportional election.
}
\examples{
fokus::election_tickets(ballot_date = "2019-10-20",
                        lvl = "federal",
                        canton = "aargau")
}
\seealso{
Other election predicate functions:
\code{\link{election_candidate_string}()},
\code{\link{election_candidates}()},
\code{\link{election_names_combined}()},
\code{\link{election_name}()},
\code{\link{election_parties}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
