% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{emphasize}
\alias{emphasize}
\title{Emphasize xth element of character vector (Markdown)}
\usage{
emphasize(x, which = TRUE, emph = "**")
}
\arguments{
\item{x}{The input as a character vector.}

\item{which}{The indices of the elements to be emphasized.}

\item{emph}{The character sequence used for emphasis.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Emphasize xth element of character vector (Markdown)
}
