% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{expand_q_tibble}
\alias{expand_q_tibble}
\title{Expand questionnaire tibble to long format}
\usage{
expand_q_tibble(q_tibble)
}
\arguments{
\item{q_tibble}{Questionnaire tibble as returned by \code{\link[=gen_q_tibble]{gen_q_tibble()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Expands a \link[=gen_q_tibble]{questionnaire tibble} to \href{https://en.wikipedia.org/wiki/Wide_and_narrow_data}{long format}.
}
\seealso{
Other questionnaire generation functions:
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_q}()},
\code{\link{gen_q_md}()},
\code{\link{gen_q_tibble}()},
\code{\link{qx}}
}
\concept{q_gen}
\keyword{internal}
