% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{expand_qstnr_tibble}
\alias{expand_qstnr_tibble}
\title{Expand questionnaire tibble to long format}
\usage{
expand_qstnr_tibble(qstnr_tibble)
}
\arguments{
\item{qstnr_tibble}{Questionnaire tibble as returned by \code{\link[=gen_qstnr_tibble]{gen_qstnr_tibble()}}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Expands a \link[=gen_qstnr_tibble]{questionnaire tibble} to \href{https://en.wikipedia.org/wiki/Wide_and_narrow_data}{long format}.
}
\seealso{
Other questionnaire generation functions:
\code{\link{gen_qstnr_md}()},
\code{\link{gen_qstnr_tibble}()}
}
\concept{qstnr_gen}
\keyword{internal}
