% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{export_print_recipients}
\alias{export_print_recipients}
\title{Export print recipients data}
\usage{
export_print_recipients(ballot_date = ballot_dates, canton = cantons)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
\code{NULL} if no export for the specified ballot date is possible, otherwise a \link[tibble:tbl_df-class]{tibble} of the exported data, invisibly.
}
\description{
Exports a CSV dataset containing the two variables \code{id_statistical_office} and \code{receives_print} to the \link[=print_fokus_private_structure]{private FOKUS directory}.
}
\seealso{
Other questionnaire generation functions
\code{\link{expand_q_tibble}()},
\code{\link{export_q}()},
\code{\link{gen_q_md}()},
\code{\link{gen_q_tibble}()}
}
\concept{q_gen}
