% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{export_q}
\alias{export_q}
\title{Export questionnaire data}
\usage{
export_q(
  ballot_date = all_ballot_dates,
  canton = cantons(ballot_date),
  verbose = FALSE,
  incl_csv = TRUE,
  incl_html = TRUE,
  incl_xlsx = incl_html,
  deploy = TRUE,
  local_deploy_path = getOption("fokus.q.local_deploy_path"),
  upload_to_g_drive = TRUE,
  g_drive_folder = "fokus/aargau/Umfragen/Dateien für Umfrageinstitut/Fragebogen/"
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{verbose}{Whether or not to print detailed progress information during questionnaire generation and Google Drive file upload. Note that questionnaire
generation takes considerably more time when this is set to \code{TRUE}.}

\item{incl_csv}{Whether or not to also generate and export a CSV version of the questionnaire.}

\item{incl_html}{Whether or not to also generate and export an HTML version of the questionnaire.}

\item{incl_xlsx}{Whether or not to also generate and export an XLSX version of the questionnaire.}

\item{deploy}{Whether or not to deploy the generated files as a static site to the Git repository specified under \code{local_deploy_path}.}

\item{local_deploy_path}{Local filesystem path to the Git repository of the static site to deploy the generated files files to. Ignored if \code{deploy = FALSE}.}

\item{upload_to_g_drive}{Whether or not to upload the generated files to the Google Drive folder \code{g_drive_folder}.}

\item{g_drive_folder}{Google Drive folder to deploy the generated files to. Ignored if \code{upload_to_g_drive = FALSE}.}
}
\description{
Generates the \link[=gen_q_tibble]{questionnaire tibble}, the \link[=gen_q_md]{Markdown questionnaire} and optionally a CSV, an HTML and an XLSX version of it, and
writes all of them to the \link[=print_fokus_private_structure]{private FOKUS directory} and optionally deploys them as a static site (\code{local_deploy_path}) and
uploads them to a Google Drive folder (\code{g_drive_folder}).
}
\seealso{
Other questionnaire and survey functions and datasets:
\code{\link{elections}},
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_q_all}()},
\code{\link{export_qr_codes}()},
\code{\link{postal_dispatch_string}()},
\code{\link{proposals}},
\code{\link{qx}}
}
\concept{q_survey}
