% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{export_q_all}
\alias{export_q_all}
\title{Export all questionnaires}
\usage{
export_q_all(incl_xlsx = FALSE, upload_to_g_drive = FALSE, deploy = FALSE)
}
\arguments{
\item{incl_xlsx}{Whether or not to also generate and export an XLSX version of the questionnaire.}

\item{upload_to_g_drive}{Whether or not to upload the generated files to the Google Drive folder \code{g_drive_folder}.}

\item{deploy}{Whether or not to deploy the generated files as a static site to the Git repository specified under \code{local_deploy_path}.}
}
\description{
Exports \emph{all} questionnaires, \emph{softly} by default (i.e. without an XLSX version and deploying/uploading).
}
\details{
Useful to test and inspect latest changes in generated questionnaire files.
}
\keyword{internal}
