% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{export_qstnr}
\alias{export_qstnr}
\title{Export questionnaire data}
\usage{
export_qstnr(
  qstnr_tibble,
  path,
  verbose = FALSE,
  incl_csv = TRUE,
  incl_html = TRUE,
  incl_xlsx = incl_html,
  upload_to_g_drive = FALSE,
  g_drive_folder = "fokus/aargau/Umfragen/Dateien für Umfrageinstitut/Fragebogen/"
)
}
\arguments{
\item{qstnr_tibble}{Questionnaire tibble as returned by \code{\link[=gen_qstnr_tibble]{gen_qstnr_tibble()}}.}

\item{path}{Path to the directory to write the generated questionnaire files to. A character scalar.}

\item{verbose}{Whether or not to print detailed progress information during questionnaire generation and Google Drive file upload. Note that questionnaire
generation takes considerably more time when this is set to \code{TRUE}.}

\item{incl_csv}{Whether or not to also generate and export a CSV version of the questionnaire.}

\item{incl_html}{Whether or not to also generate and export an HTML version of the questionnaire.}

\item{incl_xlsx}{Whether or not to also generate and export an XLSX version of the questionnaire.}

\item{upload_to_g_drive}{Whether or not to upload the generated files to the Google Drive folder \code{g_drive_folder}.}

\item{g_drive_folder}{Google Drive folder to upload the generated files to. Ignored if \code{upload_to_g_drive = FALSE}.}
}
\value{
\code{path}, invisibly.
}
\description{
Generates the \link[=gen_qstnr_tibble]{questionnaire tibble}, the \link[=gen_qstnr_md]{Markdown questionnaire} and optionally a CSV, an HTML and an XLSX version of it,
and writes all of them to \code{path} and optionally uploads them to a Google Drive folder.
}
\details{
The generated files are named according to the scheme \verb{\{ballot_date\}_\{canton\}.\{ext\}}, so if the \code{qstnr_tibble} of the 2018-09-23 survey in the canton of
Aargau is input, the following files will be written to \code{path} by default:
\itemize{
\item \verb{2018-09-23_aargau.csv}
\item \verb{2018-09-23_aargau.html}
\item \verb{2018-09-23_aargau.md}
\item \verb{2018-09-23_aargau.xlsx}
}
}
\seealso{
Other data export functions:
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_qr_codes}()},
\code{\link{write_private_file}()}
}
\concept{data_export}
