% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{export_survey_data}
\alias{export_survey_data}
\title{Export generated survey data}
\usage{
export_survey_data(
  data,
  lang = NULL,
  merged = NULL,
  auth_token = pal::pkg_config_val("token_repo_private")
)
}
\arguments{
\item{data}{FOKUS survey dataset. A \link[tibble:tbl_df-class]{tibble}.}

\item{lang}{Language. Either \code{NULL} to read the language from \code{data}'s \code{fokus_lang} attribute or one of
\code{"de"} or \code{"en"}.}

\item{merged}{Whether or not \code{data} is a \emph{merged} dataset, i.e. one that includes the data from \strong{all} ballot dates that were covered by FOKUS surveys up
until the most recent \code{ballot_date} contained in \code{data}. See section \emph{Merged data} below for details.

If \code{NULL}, a merged dataset is assumed if \code{data} covers multiple \code{ballot_date}s.}

\item{auth_token}{\href{https://gitlab.com/-/user_settings/personal_access_tokens}{Personal access token of a gitlab.com account} with access to the \href{https://gitlab.com/zdaarau/private/fokus_private/}{private FOKUS repository}.}
}
\value{
\code{data}, invisibly.
}
\description{
Exports a FOKUS survey dataset generated via \emph{TODO} to the \link[=print_private_repo_structure]{private FOKUS repository}.
}
\section{Merged data}{
Merged datasets differ from single-ballot-date datasets in two ways:
\enumerate{
\item Column \strong{labels} (the \code{label} attribute) are
ballot-date-independent, i.e. do not include proposal or election
names, and in case of question rewordings over time, the very
\emph{latest} question wording of all FOKUS surveys in that canton is
used (which might have been introduced only after \code{ballot_date}). So
be aware that the questions in the labels do not necessarily
correspond to the actual questions asked in the FOKUS survey for
\code{ballot_date}.
\item Variables whose \strong{factor levels} vary across ballot dates, such as
those capturing skill questions, are converted to type character. Be
aware that these variables usually aren’t really comparable over
time.
}
}

\seealso{
Other data export functions:
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_qr_codes}()},
\code{\link{export_qstnr}()},
\code{\link{write_private_file}()}
}
\concept{data_export}
