% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-package.R
\docType{package}
\name{fokus-package}
\alias{fokus}
\alias{fokus-package}
\title{fokus: Provides API-like functions around the FOKUS Aargau
    Research Project of the Centre for Democracy Studies Aarau (ZDA)}
\description{
fokus provides functions to (pre-)process and
    clean the raw data, analyze and visualize the cleaned data and create
    the questionnaires of the FOKUS Aargau surveys as well as other
    utility functions around the whole research project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/salim_b/r/pkgs/fokus}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/fokus/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
