% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{gen_pkg_data}
\alias{gen_pkg_data}
\title{Generate package data}
\usage{
gen_pkg_data(data_files = c("internal", "exported", "asciicasts"))
}
\arguments{
\item{data_files}{Data \code{.Rmd} source files to execute. They are stored under the paths \verb{Rmd/data_\{data_files\}.nopurl.Rmd}.}
}
\value{
The paths to the specified \code{data_files}, invisibly.
}
\description{
Generates all the data included in this package that is generated from code not included in the \href{https://r-pkgs.org/package-structure-state.html#source-package}{source package}, i.e. outsourced to files with the \href{https://rpkg.dev/pkgpurl/reference/purl_rmd.html#-rmd-files-excluded-from-purling}{\code{.nopurl} suffix in their filenames}.
}
\keyword{internal}
