% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{gen_q}
\alias{gen_q}
\title{Generate questionnaire}
\usage{
gen_q(ballot_date = ballot_dates, canton = cantons, quiet = FALSE)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{quiet}{Whether or not to suppress printing status output from internal processing.

Either as a character or a \link[base:Dates]{date} scalar.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This is the main questionnaire generation interface that's doing all the magic. This function is only called for its side effects, it doesn't return
anything.
}
\seealso{
\code{\link{qx_suppl}}

Other questionnaire generation functions
\code{\link{pick_right}()},
\code{\link{q_md}()},
\code{\link{q_tibble}()}
}
\concept{q_gen}
