% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{gen_q_md}
\alias{gen_q_md}
\title{Generate Markdown questionnaire}
\usage{
gen_q_md(q_tibble)
}
\value{
A character vector.
}
\description{
Generate Markdown questionnaire
}
\seealso{
Other questionnaire generation functions
\code{\link{gen_q_tibble}()},
\code{\link{gen_q}()}
}
\concept{q_gen}
\keyword{internal}
