% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{gen_qstnr_md}
\alias{gen_qstnr_md}
\title{Generate Markdown questionnaire}
\usage{
gen_qstnr_md(qstnr_tibble, incl_title = FALSE)
}
\arguments{
\item{qstnr_tibble}{Questionnaire tibble as returned by \code{\link[=gen_qstnr_tibble]{gen_qstnr_tibble()}}.}

\item{incl_title}{Whether or not to generate an \verb{<h1>} questionnaire title at the beginning of the document. If the result is intended to be fed to Pandoc,
it's recommended to set this to \code{FALSE} and provide the title via \href{https://pandoc.org/MANUAL.html#option--metadata}{Pandoc's \code{--metadata} option} instead.}
}
\value{
A character vector.
}
\description{
Generate Markdown questionnaire
}
\seealso{
Other questionnaire generation functions:
\code{\link{expand_qstnr_tibble}()},
\code{\link{gen_qstnr_tibble}()}
}
\concept{qstnr_gen}
\keyword{internal}
