% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{gen_qstnr_tibble}
\alias{gen_qstnr_tibble}
\title{Generate questionnaire tibble}
\usage{
gen_qstnr_tibble(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  verbose = FALSE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{verbose}{Whether or not to print detailed progress information during questionnaire generation. Note that it will take considerably more time when this
is set to \code{TRUE}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Generate questionnaire tibble
}
\seealso{
Other questionnaire generation functions:
\code{\link{expand_qstnr_tibble}()},
\code{\link{gen_qstnr_md}()}
}
\concept{qstnr_gen}
\keyword{internal}
