% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_ballot_type}
\alias{has_ballot_type}
\title{Determine whether ballot includes type}
\usage{
has_ballot_type(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  ballot_type = ballot_types(ballot_date = ballot_date, lvls = lvls, canton = canton),
  prcds = all_prcds
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvls} includes \code{"cantonal"} or \code{ballot_type = "election"}.}

\item{ballot_type}{Ballot type. One of \code{"referendum"} or \code{"election"}.}

\item{prcds}{Election procedure(s). One or more of \code{"proportional"} and \code{"majoritarian"}.

Only relevant if \code{ballot_type = "election"}.}
}
\value{
A logical vector of the same length as \code{lvl} (× \code{prcds}), named after \verb{lvls(.prcds).ballot_type}.
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) covered the specified
\link[=ballot_types]{ballot types}.
}
\examples{
fokus::has_ballot_type(ballot_date = "2018-09-23",
                       canton = "aargau")
                        
fokus::has_ballot_type(ballot_date = "2018-09-23",
                       canton = "aargau",
                       ballot_type = "election")
                        
fokus::has_ballot_type(ballot_date = "2019-10-20",
                       canton = "aargau",
                       ballot_type = "election",
                       prcds = "proportional")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{combos_ballot_types}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{is_representative}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()},
\code{\link{survey_channels}()}
}
\concept{predicate_fundamental}
