% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_ballot_types}
\alias{has_ballot_types}
\title{Determine whether ballot types are included}
\usage{
has_ballot_types(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = all_cantons,
  ballot_types = all_ballot_types
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{ballot_types}{Ballot type(s). One or more of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}}
}
\value{
A logical vector of the same length as \code{ballot_types}.
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) includes the specified
\link[=ballot_types]{ballot types}.
}
\examples{
fokus::has_ballot_types(ballot_date = "2018-09-23",
                        canton = "aargau")
                        
fokus::has_ballot_types(ballot_date = "2018-09-23",
                        canton = "aargau",
                        ballot_types = "election")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()}
}
\concept{predicate_fundamental}
