% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_election_nrs}
\alias{has_election_nrs}
\title{Determine whether ballot includes elections}
\usage{
has_election_nrs(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  prcds = all_prcds,
  election_nrs = NULL
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{prcds}{Election procedure(s). One or more of
\itemize{
\item \code{"proportional"}
\item \code{"majoritarian"}
}}

\item{election_nrs}{Election number(s) to test for. An integerish vector or \code{NULL}. If \code{NULL}, falls back to \code{\link[=n_elections]{1:n_elections()}} present
matching the specified parameters.}
}
\value{
Either
\itemize{
\item \strong{an unnamed logical scalar} if \code{election_nrs} is \code{NULL} and there was no FOKUS-covered election of any \code{prcds} at \code{ballot_date} on any \code{lvls}, or
\item \strong{a logical vector} of the same length as \code{lvls} × \code{prcds} × \code{election_nrs}, \strong{named} after \code{lvls.prcds.election_nrs}.
}
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political levels covered the specified
election numbers of the specified procedures.
}
\examples{
fokus::has_election_nrs(ballot_date = "2019-10-20",
                        canton = "aargau")

fokus::has_election_nrs(ballot_date = "2019-10-20",
                        lvls = "federal",
                        canton = "aargau",
                        prcds = "proportional")
                        
fokus::has_election_nrs(ballot_date = "2019-10-20",
                        canton = "aargau",
                        election_nrs = 1:2)

fokus::has_election_nrs(ballot_date = "2018-09-23",
                        canton = "aargau")

fokus::has_election_nrs(ballot_date = "2020-10-18",
                        lvls = "federal",
                        canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()}
}
\concept{predicate_fundamental}
