% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_lvl}
\alias{has_lvl}
\title{Determine whether ballot includes a political level}
\usage{
has_lvl(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{The political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A logical scalar.
}
\description{
Determines whether or not the ballot in the specified canton at the specified date includes the specified political level.
}
\examples{
fokus::has_lvl(ballot_date = "2018-09-23",
               lvl = "federal",
               canton = "aargau")
}
\seealso{
Other predicate functions
\code{\link{ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_referendum}()},
\code{\link{n_election_candidates}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{n_skill_questions}()},
\code{\link{skill_question_answer_nr}()}
}
\concept{predicate}
