% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_proposal_arguments}
\alias{has_proposal_arguments}
\title{Determine whether arguments have been queried for referendum proposal}
\usage{
has_proposal_arguments(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "referendum"),
  canton = cantons(ballot_date),
  proposal_nr = 1L,
  sides = all_argument_sides
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integerish scalar.}

\item{sides}{Proposal argument side(s). One or more of \code{"pro"} and \code{"contra"}.}
}
\value{
A logical scalar.
}
\description{
Determines whether or not arguments have been queried for the specified referendum proposal.
}
\examples{
fokus::has_proposal_arguments(ballot_date = "2023-06-18",
                              lvl = "federal",
                              canton = "aargau")

fokus::has_proposal_arguments(ballot_date = "2023-06-18",
                              lvl = "federal",
                              canton = "aargau",
                              proposal_nr = 2)
}
\seealso{
Other proposal predicate functions:
\code{\link{argument_proposal_nrs}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()},
\code{\link{has_proposal_main_motives}()},
\code{\link{is_proposal_name_plural}()},
\code{\link{main_motive_proposal_nrs}()},
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
