% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_proposal_nrs}
\alias{has_proposal_nrs}
\title{Determine whether ballot includes proposals}
\usage{
has_proposal_nrs(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = all_cantons,
  proposal_nrs = NULL
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{proposal_nrs}{Proposals number(s) to check. An integerish vector. If \code{NULL}, falls back to the \link[=n_proposals]{number of proposals} present matching the
specified parameters.}
}
\value{
A logical vector of the same length as the number of proposals present matching the specified parameters (\code{lvls} × \code{proposal_nrs}).
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political levels covered the specified
proposal numbers.
}
\examples{
fokus::has_proposal_nrs(ballot_date = "2018-09-23",
                        canton = "aargau")
                        
fokus::has_proposal_nrs(ballot_date = "2018-09-23",
                        canton = "aargau",
                        proposal_nrs = 1:5)
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_ballot_types}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()}
}
\concept{predicate_fundamental}
