% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{is_election}
\alias{is_election}
\title{Determine if ballot type includes an election}
\usage{
is_election(canton = cantons, ballot_date = as.character(ballot_dates))
}
\arguments{
\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"} )`
}}

\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"} )`
}}
}
\value{
A logical scalar.
}
\description{
Determine if ballot type includes an election
}
\seealso{
Other fundamental predicates
\code{\link{ballot_type}()},
\code{\link{is_referendum}()}
}
\concept{fundamental}
