% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{is_referendum}
\alias{is_referendum}
\title{Determine if ballot type includes a referendum}
\usage{
is_referendum(canton = cantons, ballot_date = ballot_dates)
}
\arguments{
\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}
}
\value{
A logical scalar.
}
\description{
Determine if ballot type includes a referendum
}
\seealso{
Other fundamental predicates
\code{\link{ballot_type}()},
\code{\link{is_election}()}
}
\concept{fundamental}
