% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{is_skill_question}
\alias{is_skill_question}
\title{Determine whether variable is skill question}
\usage{
is_skill_question(v_names)
}
\arguments{
\item{v_names}{A character vector of FOKUS variable names.}
}
\value{
A logical vector.
}
\description{
Determines whether or not a variable is a skill question. Vectorized.
}
\seealso{
Other variable-related functions
\code{\link{shorten_v_names}()},
\code{\link{v_lvls}()}
}
\concept{variable}
