% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{is_skill_question_v}
\alias{is_skill_question_v}
\title{Determine whether variable is skill question}
\usage{
is_skill_question_v(v_names)
}
\arguments{
\item{v_names}{A character vector of FOKUS variable names.}
}
\value{
A logical vector of the same length as \code{v_names}.
}
\description{
Determines for each variable whether or not it is a skill question.
}
\details{
Note that the determination is performed by simply parsing \code{v_names}.
}
\seealso{
Other variable-related functions:
\code{\link{shorten_v_names}()},
\code{\link{v_lvls}()},
\code{\link{v_proposal_nr}()},
\code{\link{v_skill_question_nr}()}
}
\concept{variable}
