% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{lang_to_locale}
\alias{lang_to_locale}
\title{Convert language code to country-specific locale ID}
\usage{
lang_to_locale(lang = c("de", "en"))
}
\arguments{
\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character scalar.
}
\description{
Converts a language code as used in many of this package's functions to a country-specific locale identifier.
}
\examples{
fokus:::lang_to_locale("de")
fokus:::lang_to_locale("en")
}
\keyword{internal}
