% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{lgl_to_unicode}
\alias{lgl_to_unicode}
\title{Convert logical vector to Unicode symbols ✅ and ❌}
\usage{
lgl_to_unicode(x)
}
\arguments{
\item{x}{A logical vector.}
}
\value{
A character vector.
}
\description{
Convert logical vector to Unicode symbols ✅ and ❌
}
\examples{
fokus::lgl_to_unicode(c(TRUE, TRUE, FALSE, NA))
}
