% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{lvl_prose}
\alias{lvl_prose}
\title{Transform political level into prose string}
\usage{
lvl_prose(lvl, lang = all_langs)
}
\arguments{
\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=all_lvls]{political level} into a full prose string representation.
}
\examples{
fokus::all_lvls |>
  purrr::map(\(x) tibble::tibble(lvl = x,
                                 de = fokus::lvl_prose(lvl = x,
                                                       lang = "de"),
                                 en = fokus::lvl_prose(lvl = x,
                                                       lang = "en"))) |>
  purrr::list_rbind()
}
\seealso{
Other prose text functions:
\code{\link{argument_side_prose}()},
\code{\link{ballot_type_prose}()},
\code{\link{motive_type_prose}()},
\code{\link{postal_dispatch_type_prose}()},
\code{\link{postal_dispatch_way_prose}()},
\code{\link{proposal_name_declined_de}()}
}
\concept{prose}
