% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{lvls}
\alias{lvls}
\title{Determine ballot's political levels}
\usage{
lvls(
  ballot_date = all_ballot_dates,
  canton = all_cantons,
  ballot_types = all_ballot_types
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{ballot_types}{Ballot type(s). One or more of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}}
}
\value{
A character vector of \link[=all_lvls]{political levels}.
}
\description{
Determines the political levels covered by the FOKUS survey for the specified canton at the specified ballot date.
}
\examples{
fokus::lvls(ballot_date = "2018-09-23",
            canton = "aargau")

fokus::lvls(ballot_date = "2018-09-23",
            canton = "aargau",
            ballot_types = "election")
}
