% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{main_motive_proposal_nrs}
\alias{main_motive_proposal_nrs}
\title{Get referendum proposal numbers with arguments}
\usage{
main_motive_proposal_nrs(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "referendum"),
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}
}
\value{
An integer vector.
}
\description{
Determines the referendum proposal numbers covered by the FOKUS survey for the specified canton at the specified ballot date on the specified political
level for which at least one pro/contra argument has been queried.
}
\examples{
fokus::main_motive_proposal_nrs(ballot_date = "2023-06-18",
                                lvl = "federal",
                                canton = "aargau")
}
\seealso{
Other proposal predicate functions:
\code{\link{argument_proposal_nrs}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()},
\code{\link{has_proposal_arguments}()},
\code{\link{has_proposal_main_motives}()},
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_name}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
