% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{motive_type_prose}
\alias{motive_type_prose}
\title{Transform motive type into prose string}
\usage{
motive_type_prose(type, lang = all_langs)
}
\arguments{
\item{type}{Main motive type. One of \code{"yes"} or \code{"no"}.}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=proposal_main_motives]{proposal main motives} type into a full prose string representation.
}
\examples{
fokus::all_main_motive_types |>
  purrr::map(\(x) tibble::tibble(type = x,
                                 de = fokus::motive_type_prose(type = x,
                                                               lang = "de"),
                                 en = fokus::motive_type_prose(type = x,
                                                               lang = "en"))) |>
  purrr::list_rbind()
}
\seealso{
Other prose text functions:
\code{\link{argument_side_prose}()},
\code{\link{ballot_type_prose}()},
\code{\link{postal_dispatch_type_prose}()},
\code{\link{postal_dispatch_way_prose}()},
\code{\link{proposal_name_declined_de}()}
}
\concept{prose}
