% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_proposal_arguments}
\alias{n_proposal_arguments}
\title{Get proposal's number of arguments}
\usage{
n_proposal_arguments(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  proposal_nr = 1L
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level \link[=lvls]{covered at} \code{ballot_date} and \code{canton}.}

\item{canton}{Canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}. Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integer scalar.}
}
\value{
An integer scalar.
}
\description{
Determines the number of arguments on the specified proposal.
}
\examples{
fokus::n_proposal_arguments(ballot_date = "2018-09-23",
                            lvl = "cantonal",
                            canton = "aargau",
                            proposal_nr = 1)
}
\seealso{
Other proposal predicate functions:
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_name}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
