% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_proposals}
\alias{n_proposals}
\title{Get number of referendum proposals}
\usage{
n_proposals(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
An integer.
}
\description{
Determines the number of referendum proposals covered by the FOKUS survey for the specified canton at the specified ballot date on the specified political
level(s).
}
\details{
The number of \emph{federal} proposals is independent from the canton, i.e. the returned number of \emph{federal} proposals at a specific ballot date is always the
same, thus \code{canton} is ignored if \code{!("cantonal" \%in\% lvls)}.
}
\examples{
fokus::n_proposals(ballot_date = "2018-09-23",
                   canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_ballot_types}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{prcds}()}
}
\concept{predicate_fundamental}
