% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{opts}
\alias{opts}
\title{List package-specific options}
\usage{
opts(pretty_colnames = FALSE)
}
\arguments{
\item{pretty_colnames}{Whether or not to return prose colnames. If \code{FALSE}, snake-cased colnames safe to use in R expressions are returned.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a tibble of R options specific to the FOKUS package.
}
\examples{
fokus:::opts()
}
\keyword{internal}
