% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{path_private}
\alias{path_private}
\title{Assemble a private FOKUS directory path}
\usage{
path_private(...)
}
\arguments{
\item{...}{Filesystem path elements as unnamed vectors.}
}
\value{
A \link[fs:fs_path]{path}.
}
\description{
Assembles a path from the private FOKUS directory root plus any additional path elements supplied.
}
\details{
See \code{\link[=print_fokus_private_structure]{print_fokus_private_structure()}} for details about the private FOKUS directory.
}
\keyword{internal}
