% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{phrase_date}
\alias{phrase_date}
\title{Phrase (ballot) date}
\usage{
phrase_date(
  x = pal::pkg_config_val("ballot_date"),
  format = "date_long",
  lang = pal::pkg_config_val("lang")
)
}
\arguments{
\item{x}{Date(s) or datetime(s) to format. A \link[base:Dates]{date}, \link[base:DateTimeClasses]{datetime}, factor or character vector.}

\item{format}{Format string(s) defining how to phrase \code{x}. Possible values include
\code{"date_short"}, \code{"date_medium"}, \code{"date_long"}, \code{"date_full"}, \code{"time_short"}, \code{"time_medium"}, \code{"time_long"} and \code{"time_full"}.
See \code{\link[stringi:stri_datetime_format]{stringi::stri_datetime_format()}} for details.}

\item{lang}{Language. One of \code{"de"}.}
}
\value{
A character scalar.
}
\description{
Phrases a date according to the specified \code{format} and \code{lang}.
}
\examples{
fokus::phrase_date()
}
\seealso{
Other text phrasing functions:
\code{\link{phrase}()},
\code{\link{phrase_election_candidate}()},
\code{\link{phrase_proposal_name_de}()}
}
\concept{phrase}
