% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{phrase_election_candidate}
\alias{phrase_election_candidate}
\title{Phrase majoritarian election's candidate(s)}
\usage{
phrase_election_candidate(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "election"),
  canton = cantons(ballot_date),
  election_nr = 1L,
  candidate_nrs = NULL,
  incl_party = TRUE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{Election number. A positive integerish scalar.}

\item{candidate_nrs}{Election candidate numbers to include. A vector of positive integers or \code{NULL}. If \code{NULL}, all candidates will be included.}

\item{incl_party}{Whether or not to include the candidate's political party in the resulting string (in parentheses).}
}
\value{
A character vector.
}
\description{
Assembles one or more majoritarian election candidate strings consisting of the candidate's first name, last name and optionally political party (in
parentheses).
}
\examples{
fokus::phrase_election_candidate(ballot_date = "2019-10-20",
                                 lvl = "cantonal",
                                 canton = "aargau",
                                 candidate_nrs = 1:3)
}
\seealso{
Other text phrasing functions:
\code{\link{phrase_proposal_name_de}()},
\code{\link{phrase}()}
}
\concept{phrase}
