% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{phrase_proposal_name_de}
\alias{phrase_proposal_name_de}
\title{Get declined German proposal name}
\usage{
phrase_proposal_name_de(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "referendum"),
  canton = cantons(ballot_date),
  proposal_nr = 1L,
  type = all_name_types,
  case = c("nominative", "genitive", "dative", "accusative")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integerish scalar.}

\item{type}{Name type. One of \code{"short"} or \code{"long"}.}

\item{case}{The grammatical case of the \code{noun}.}
}
\value{
A character scalar.
}
\description{
Declines a German \link[=proposal_name]{proposal name} to the specified grammatical \code{case}. Simple convenience function combining \code{\link[=proposal_name]{proposal_name()}},
\code{\link[=proposal_name_gender]{proposal_name_gender()}} and \code{\link[salim:decline_noun_de]{salim::decline_noun_de()}}.
}
\examples{
fokus::phrase_proposal_name_de(ballot_date = "2023-06-18",
                               lvl = "cantonal",
                               canton = "aargau",
                               proposal_nr = 1,
                               type = "short",
                               case = "dative")

fokus::phrase_proposal_name_de(ballot_date = "2023-06-18",
                               lvl = "cantonal",
                               canton = "aargau",
                               proposal_nr = 1,
                               type = "short",
                               case = "genitive")
}
\seealso{
Other text phrasing functions:
\code{\link{phrase}()},
\code{\link{phrase_election_candidate}()}
}
\concept{phrase}
