% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{pick_right}
\alias{pick_right}
\title{Pick the right value of a certain questionnaire key based on canton and ballot date (recursively)}
\usage{
pick_right(l, canton, ballot_date)
}
\arguments{
\item{l}{The questionnaire subitem. A list object.}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}

Either as a character or a \link[base:Dates]{date} scalar.}
}
\value{
The value of \code{l} that corresponds to \code{canton} and \code{ballot_date}.
}
\description{
Pick the right value of a certain questionnaire key based on canton and ballot date (recursively)
}
\seealso{
Other questionnaire-related functions
\code{\link{gen_q}()},
\code{\link{q_md}()},
\code{\link{q_tibble}()},
\code{\link{read_q_supplemental}()},
\code{\link{read_q}()}
}
\concept{questionnaire}
