% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{political_issues}
\alias{political_issues}
\title{Get political issues}
\usage{
political_issues(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lang = c("de", "en")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character vector.
}
\description{
Returns the political issues in the specified language at the specified ballot date.
}
\examples{
fokus::political_issues(ballot_date = "2019-10-20",
                        lang = "en")
}
\seealso{
Other predicate functions:
\code{\link{ballot_title}()},
\code{\link{postal_dispatch_way}()},
\code{\link{response_options}()}
}
\concept{predicate_other}
