% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{postal_dispatch_string}
\alias{postal_dispatch_string}
\title{Transform postal dispatch way into prose string}
\usage{
postal_dispatch_string(
  dispatch_way = all_postal_dispatch_ways,
  lang = c("de", "en")
)
}
\arguments{
\item{dispatch_way}{Postal dispatch way One of
\itemize{
\item \code{"A"}
\item \code{"B"}
\item \code{"B bulk mailing"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=postal_dispatch_way]{postal dispatch way} into a full prose string representation.
}
\examples{
fokus::postal_dispatch_way(ballot_date = "2018-09-23",
                           canton = "aargau",
                           dispatch_type = "invitation") |>
  fokus::postal_dispatch_string()
}
\seealso{
Other questionnaire and survey functions and datasets:
\code{\link{elections}},
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_q_all}()},
\code{\link{export_qr_codes}()},
\code{\link{export_q}()},
\code{\link{proposals}},
\code{\link{qx}}
}
\concept{q_survey}
