% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{postal_dispatch_way}
\alias{postal_dispatch_way}
\title{Get postal dispatch way}
\usage{
postal_dispatch_way(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  dispatch_type = all_postal_dispatch_types
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{dispatch_type}{Postal dispatch type. One of \code{"invitation"}, \code{"prepaid_reply_envelope"} or \code{"reminder"}.}
}
\value{
A character scalar.
}
\description{
Returns the FOKUS survey's postal dispatch way of the specified type for the specified canton at the specified ballot date.
}
\examples{
fokus::postal_dispatch_way(ballot_date = "2018-09-23",
                           canton = "aargau",
                           dispatch_type = "invitation")
}
\seealso{
Other predicate functions:
\code{\link{ballot_title}()},
\code{\link{political_issues}()},
\code{\link{response_options}()}
}
\concept{postal_dispatch}
\concept{predicate_other}
