% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{prcds}
\alias{prcds}
\title{Get ballot's election procedures}
\usage{
prcds(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}
}
\value{
A character vector of \link[=all_prcds]{election procedures}.
}
\description{
Determines the election procedures covered by the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s).
}
\examples{
fokus::prcds(ballot_date = "2019-10-20",
             canton = "aargau")

fokus::prcds(ballot_date = "2019-10-20",
             lvls = "cantonal",
             canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_ballot_types}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()}
}
\concept{predicate_fundamental}
