% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{print_fokus_private_structure}
\alias{print_fokus_private_structure}
\title{Print expected structure of the private FOKUS directory}
\usage{
print_fokus_private_structure()
}
\value{
A character scalar.
}
\description{
Returns a textual representation of the expected structure of the private FOKUS directory, formatted as a Markdown \href{https://pandoc.org/MANUAL.html#extension-fenced_code_blocks}{fenced code block}.
}
\details{
For part of this package’s functionality, a separate private directory
(defined by the R option \code{fokus.path_private}) is required which is
expected to be organized according to a very specific structure. This
private directory i.a. must contain additional sensitive, non-public
(survey) data under the \verb{data/} subdirectory and certain files are
written to its \verb{output/} subdirectory.
}
\keyword{internal}
