% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{print_opts}
\alias{print_opts}
\title{Pretty-print package-specific options}
\usage{
print_opts()
}
\value{
A character vector.
}
\description{
Pretty-prints the output of \code{\link[=opts]{opts()}} as a Markdown \link[pal:pipe_table]{pipe table}.
}
\examples{
fokus:::print_opts()
}
\keyword{internal}
