% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{print_private_repo_structure}
\alias{print_private_repo_structure}
\title{Print structure of the private FOKUS repository}
\usage{
print_private_repo_structure()
}
\value{
A character scalar.
}
\description{
Returns a textual representation of the structure of the private FOKUS repository, formatted as a Markdown \href{https://pandoc.org/MANUAL.html#extension-fenced_code_blocks}{fenced code block}.
}
\details{
For part of this package’s functionality, a \href{https://gitlab.com/-/user_settings/personal_access_tokens}{personal access token (PAT) of a gitlab.com account} with
access to the \href{https://gitlab.com/zdaarau/private/fokus_private/}{private FOKUS repository} is
required. This repository contains additional sensitive, non-public
(survey) data under the \verb{raw/} subdirectory and certain files are
written to its \verb{generated/} subdirectory.
}
\seealso{
Other private FOKUS repository functions:
\code{\link{read_private_file}()},
\code{\link{url_repo_private}()},
\code{\link{write_private_file}()}
}
\concept{private}
\keyword{internal}
