% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{print_wd_structure}
\alias{print_wd_structure}
\title{Print working directory structure}
\usage{
print_wd_structure()
}
\value{
A character scalar.
}
\description{
For part of this package's functionality, a specific structure of the working directory (defined by the R option \code{fokus.path_wd}) is expected (input) or enforced (output). This function returns a textual representation of this structure, formatted as a Markdown \href{https://pandoc.org/MANUAL.html#extension-fenced_code_blocks}{fenced code block}.
}
