% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{proposal_name}
\alias{proposal_name}
\title{Get proposal name}
\usage{
proposal_name(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  proposal_nr = 1L,
  lang = c("de", "en"),
  type = c("short", "long")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"} .
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integer scalar.}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}

\item{type}{Name type. One of
\itemize{
\item \code{"short"}
\item \code{"long"}
}}
}
\value{
A character scalar.
}
\description{
Returns the name of the specified proposal in the specified language.
}
\examples{
fokus::proposal_name(ballot_date = "2018-09-23",
                     lvl = "cantonal",
                     canton = "aargau",
                     proposal_nr = 1,
                     type = "long")
}
\seealso{
Other proposal predicate functions:
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
