% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{proposal_name_gender}
\alias{proposal_name_gender}
\title{Get German proposal name's grammatical gender}
\usage{
proposal_name_gender(
  ballot_date = all_ballot_dates,
  lvl = all_lvls,
  canton = cantons(ballot_date),
  proposal_nr = 1L,
  type = c("short", "long")
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{proposal_nr}{Proposal number. A positive integer scalar.}

\item{type}{Name type. One of
\itemize{
\item \code{"short"}
\item \code{"long"}
}}
}
\value{
A character scalar.
}
\description{
Returns the grammatical gender of the German name of the specified proposal.
}
\examples{
fokus::proposal_name_gender(ballot_date = "2018-09-23",
                            lvl = "cantonal",
                            canton = "aargau",
                            proposal_nr = 1,
                            type = "short")
}
\seealso{
Other proposal predicate functions:
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
