% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{proposal_nrs}
\alias{proposal_nrs}
\title{Get ballot's referendum proposal numbers}
\usage{
proposal_nrs(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "referendum"),
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}
}
\value{
An integer vector.
}
\description{
Determines the referendum proposal numbers covered by the FOKUS survey for the specified canton at the specified ballot date on the specified political
level.
}
\examples{
fokus::proposal_nrs(ballot_date = "2018-09-23",
                    lvl = "cantonal",
                    canton = "aargau")
                    
fokus::proposal_nrs(ballot_date = "2018-09-23",
                    lvl = "federal",
                    canton = "aargau")

# in case of no referendum proposals, an empty integer vector is returned
fokus::proposal_nrs(ballot_date = "2019-10-20",
                    lvl = "cantonal",
                    canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()}
}
\concept{predicate_fundamental}
