% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{proposal_qstn_groups}
\alias{proposal_qstn_groups}
\title{Get referendum proposal question groups}
\usage{
proposal_qstn_groups(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = all_lvls,
  canton = cantons(ballot_date),
  proposal_nr = 1L
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integerish scalar.}
}
\value{
A character vector of question group identifiers.
}
\description{
Determines the referendum proposal \link[=all_qstn_groups]{question groups} covered by the FOKUS survey for the specified canton on the specified political level
at the specified ballot date.
}
\examples{
fokus::proposal_qstn_groups(ballot_date = "2023-06-18",
                            lvl = "federal",
                            canton = "aargau",
                            proposal_nr = 1)

fokus::proposal_qstn_groups(ballot_date = "2023-06-18",
                            lvl = "federal",
                            canton = "aargau",
                            proposal_nr = 2)
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{combos_ballot_types}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{qstn_groups_proposal_nrs}()},
\code{\link{survey_channels}()}
}
\concept{predicate_fundamental}
