% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{proposal_type}
\alias{proposal_type}
\title{Get proposal type}
\usage{
proposal_type(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton, ballot_type = "referendum"),
  canton = cantons(ballot_date),
  proposal_nr = 1L
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integerish scalar.}
}
\value{
Proposal type. One of
\itemize{
\item \code{"citizens' initiative"}
\item \code{"direct counterproposal"}
\item \code{"indirect counterproposal"}
\item \code{"mandatory referendum"}
\item \code{"optional referendum"}
}
}
\description{
Get proposal type
}
\examples{
fokus::proposal_type(ballot_date = "2018-09-23",
                     lvl = "cantonal",
                     canton = "aargau",
                     proposal_nr = 1)
}
\seealso{
Other proposal predicate functions:
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_name}()}
}
\concept{predicate_proposal}
