% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{proposal_types}
\alias{proposal_types}
\title{Referendum proposal types}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
proposal_types
}
\description{
A vector of all \link[=proposal_type]{referendum proposal types}.
}
\seealso{
\code{\link{ballot_dates}} \code{\link{response_option_types}} \code{\link{q_item_keys}}
}
\keyword{datasets}
