% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{proposals}
\alias{proposals}
\title{Proposal data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
proposals
}
\description{
A tibble containing basic referendum proposal data of all FOKUS questionnaires.
}
\examples{
fokus::proposals
}
\seealso{
Other datasets:
\code{\link{ballots}},
\code{\link{elections}},
\code{\link{qstnrs}}
}
\concept{qstnr_data}
\keyword{datasets}
