% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{q}
\alias{q}
\title{Raw FOKUS questionnaire data}
\format{
A \link[xfun:strict_list]{strict list}.
}
\usage{
q
}
\description{
A structured list of the raw questionnaire data of the FOKUS surveys.
}
\seealso{
\code{\link{qx_supplemental}} \code{\link[=gen_q]{gen_q()}} \code{\link[=q_tibble]{q_tibble()}} \code{\link[=q_md]{q_md()}} \code{\link[=pick_right]{pick_right()}}
}
\keyword{datasets}
