% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{q_item_keys}
\alias{q_item_keys}
\title{List supported questionnaire item keys}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
q_item_keys
}
\description{
A tibble of item keys supported in the \link[=raw_q]{raw FOKUS questionnaire data}.
}
\seealso{
\code{\link{ballot_dates}} \code{\link{cantons}} \code{\link{response_option_types}} \code{\link{proposal_types}}
}
\keyword{datasets}
