% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_item_val}
\alias{q_item_val}
\title{Extract questionnaire item field value}
\usage{
q_item_val(
  ballot_date = ballot_dates,
  canton = cantons,
  branch_path,
  v_name,
  key = q_item_keys$key,
  lvl = "?",
  i = NA_integer_,
  j = NA_integer_
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{branch_path}{Sequence of questionnaire table levels that lead to the \code{item} leaf node where \code{v_name} is defined. A character vector.}

\item{v_name}{Variable name, without resolved string interpolation, i.e. 1:1 as stated in the raw FOKUS questionnaire data. A character scalar.}

\item{key}{Questionnaire item key. One of
\itemize{
\item \code{"lvl"}
\item \code{"i"}
\item \code{"j"}
\item \code{"block"}
\item \code{"variable_name"}
\item \code{"who"}
\item \code{"topic"}
\item \code{"question"}
\item \code{"question_full"}
\item \code{"question_intro_lvl"}
\item \code{"question_intro_i"}
\item \code{"question_intro_j"}
\item \code{"question_common"}
\item \code{"variable_label"}
\item \code{"variable_label_common"}
\item \code{"response_options"}
\item \code{"variable_values"}
\item \code{"value_labels"}
\item \code{"value_scale"}
\item \code{"multiple_answers_allowed"}
\item \code{"randomize_response_options"}
\item \code{"is_mandatory"}
\item \code{"ballot_types"}
\item \code{"include"}
}}

\item{lvl}{Political-level loop iterator that can be referred to in field value via \link[glue:glue]{string interpolation}. A character vector.}

\item{i}{Second-level loop iterator that can be referred to in field value via \link[glue:glue]{string interpolation}. An integer vector.}

\item{j}{Third-level loop iterator that can be referred to in field value via \link[glue:glue]{string interpolation}. An integer vector.}
}
\value{
A vector of the resolved item field values. Type and length of resolved values depend on \code{key}.
}
\description{
Extracts a single questionnaire item field value for every first- for every second-level iterator from the \link[=raw_q]{raw FOKUS questionnaire data} based on the
specified variable name and branch path.
}
\details{
When used in questionnaire item field values via \link[glue:glue]{string interpolation}, be careful to not create infinite loops via circular references.
}
\examples{
fokus:::q_item_val(ballot_date = "2018-09-23",
                   canton = "aargau",
                   branch_path = c("03_proposal_specific", "004_participants"),
                   v_name = "appeal_federal_proposals",
                   key = "include")

fokus:::q_item_val(
  ballot_date = "2019-10-20",
  canton = "aargau",
  branch_path = c("02_political_start", "002_non_participation",
                  "003_election", "z01_hypothetical"),
  v_name = "hypothetical_voting_decision_federal_majoritarian_election_{i}_seat_{j}",
  key = "topic",
  i = 1:2,
  j = 1:3
)
}
\seealso{
Other questionnaire-internal functions
\code{\link{q_parties}()},
\code{\link{q_response_option_codes}()},
\code{\link{q_response_options}()}
}
\concept{q_internal}
\keyword{internal}
