% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_lbl_col_sym}
\alias{q_lbl_col_sym}
\title{Determine questionnaire data's value label column}
\usage{
q_lbl_col_sym(lang = c("de", "en"))
}
\arguments{
\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A \link[base:name]{symbol}.
}
\description{
Determines the \link[=qx]{questionnaire data} column that holds variable value labels, returned as a \link[=as.symbol]{symbol}.
}
\details{
The questionnaire data column that holds variable value labels is language-dependent. While the column
\code{response_options} holds the German value labels, their English counterpars are found in the column
\code{value_labels}. \code{q_lbl_col_sym()} is intended to ease language-agnostic questionnaire data programming.
}
\examples{
fokus:::q_lbl_col_sym(lang = "de")

fokus:::q_lbl_col_sym(lang = "en")
}
\concept{q_predicate}
\keyword{internal}
