% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_response_options}
\alias{q_response_options}
\title{Extract response options}
\usage{
q_response_options(
  type = all_response_option_types,
  lang = c("de", "en"),
  subtypes = NULL
)
}
\arguments{
\item{type}{Response option type. One of
\itemize{
\item \code{"abstain"}
\item \code{"custom"}
\item \code{"dunno"}
\item \code{"no"}
\item \code{"none"}
\item \code{"other"}
\item \code{"refusal"}
\item \code{"unspecific"}
\item \code{"yes"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}

\item{subtypes}{Hierarchical response option subtypes as a character vector, or \code{NULL}. If \code{NULL}, all subtypes are returned.}
}
\value{
A character vector.
}
\description{
Extracts response options of the specified type and optionally subtype(s) from the \link[=raw_q]{raw FOKUS questionnaire data}.
}
\details{
Note that only recurring response options are returned which are defined under the \code{response_options} top-level key in the file
\code{data-raw/questionnaire/questionnaire.toml}.
}
\examples{
fokus:::q_response_options(type = "abstain",
                           lang = "de")

fokus:::q_response_options(type = "abstain",
                           lang = "de",
                           subtypes = "election")

fokus:::q_response_options(type = "abstain",
                           lang = "de",
                           subtypes = c("election", "proportional"))
}
\seealso{
Other questionnaire-internal functions:
\code{\link{q_item_val}()},
\code{\link{q_parties}()},
\code{\link{q_response_option_codes}()}
}
\concept{q_internal}
\keyword{internal}
