% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_suppl_election}
\alias{q_suppl_election}
\title{Get raw election questionnaire data}
\usage{
q_suppl_election(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  prcd = c("proportional", "majoritarian"),
  election_nr = 1L
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{prcd}{The election procedure. One of
\itemize{
\item \code{"proportional"}
\item \code{"majoritarian"}
}}

\item{election_nr}{The election number. A positive integer scalar (in almost all cases \code{1L}).}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of raw election data from the \link[=q_suppl_lvl_canton]{canton's political-level-specific raw supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus::q_suppl_election(ballot_date = "2019-10-20",
                        lvl = "cantonal",
                        canton = "aargau",
                        prcd = "majoritarian") \%$\%
  n_seats
}
\seealso{
Raw questionnaire data \code{\link{q}} \code{\link{qx_suppl}}

Other raw questionnaire data functions
\code{\link{q_suppl_lvl_canton}()},
\code{\link{q_suppl_lvl}()},
\code{\link{q_suppl_proposal}()},
\code{\link{q_suppl_skill_questions}()},
\code{\link{q_suppl_skill_question}()},
\code{\link{q_suppl}()}
}
\concept{q_raw}
