% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_suppl_proposal}
\alias{q_suppl_proposal}
\title{Get raw proposal questionnaire data}
\usage{
q_suppl_proposal(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  proposal_nr
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{The proposal number. A positive integer scalar or \code{NULL}. If \code{NULL}, it is considered to be a non-proposal-specific skill question
(the case at elections).}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of raw proposal data from the \link[=q_suppl]{supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus::q_suppl_proposal(ballot_date = "2018-09-23",
                        lvl = "cantonal",
                        canton = "aargau",
                        proposal_nr = 1) \%$\%
  gender
}
\seealso{
Raw questionnaire data \code{\link{q}} \code{\link{qx_suppl}}

Other raw questionnaire data functions
\code{\link{q_suppl_election}()},
\code{\link{q_suppl_lvl_canton}()},
\code{\link{q_suppl_lvl}()},
\code{\link{q_suppl_skill_questions}()},
\code{\link{q_suppl_skill_question}()},
\code{\link{q_suppl}()}
}
\concept{q_raw}
