% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_tibble}
\alias{q_tibble}
\title{Generate questionnaire tibble}
\usage{
q_tibble(ballot_date = ballot_dates, canton = cantons)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Generate questionnaire tibble
}
\seealso{
\code{\link{qx_suppl}}

Other questionnaire generation functions
\code{\link{gen_q}()},
\code{\link{pick_right}()},
\code{\link{q_md}()}
}
\concept{q_gen}
