% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{qstn_groups_proposal_nrs}
\alias{qstn_groups_proposal_nrs}
\title{Get referendum proposal numbers with question group}
\usage{
qstn_groups_proposal_nrs(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = all_lvls,
  canton = cantons(ballot_date),
  qstn_groups = all_qstn_groups
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{qstn_groups}{One or more of the standardized question groups \code{"argument"}, \code{"main_motive"} or \code{"skill_question"}.}
}
\value{
An integer vector.
}
\description{
Determines the referendum proposal numbers covered by the FOKUS survey for the specified canton at the specified ballot date on the specified political
level for which at least one question of each of the specified standardized question groups has been asked.
}
\examples{
fokus::qstn_groups_proposal_nrs(ballot_date = "2023-06-18",
                                lvl = "federal",
                                canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()}
}
\concept{predicate_fundamental}
