% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{qstnr_item_keys}
\alias{qstnr_item_keys}
\title{Questionnaire item keys}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
A tibble of item keys supported in the \link[=raw_qstnr]{raw FOKUS questionnaire data}.
}
\examples{
fokus:::qstnr_item_keys
}
\seealso{
Other FOKUS metadata functions:
\code{\link{all_argument_sides}},
\code{\link{all_ballot_dates}},
\code{\link{all_ballot_types}},
\code{\link{all_cantons}},
\code{\link{all_lvls}},
\code{\link{all_postal_dispatch_types}},
\code{\link{all_postal_dispatch_ways}},
\code{\link{all_prcds}},
\code{\link{all_proposal_types}},
\code{\link{all_response_option_types}}
}
\concept{metadata}
\keyword{internal}
