% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{qstnr_lbl_col_sym}
\alias{qstnr_lbl_col_sym}
\title{Determine questionnaire data's value label column}
\usage{
qstnr_lbl_col_sym(lang = pal::pkg_config_val("lang"))
}
\arguments{
\item{lang}{Language. One of \code{"de"}.}
}
\value{
A \link[base:name]{symbol}.
}
\description{
Determines the \link[=qstnrs]{questionnaire data} column that holds variable value labels, returned as a \link[=as.symbol]{symbol}.
}
\details{
The questionnaire data column that holds variable value labels is language-dependent. While the column
\code{response_options} holds the German value labels, their English counterpars are found in the column
\code{value_labels}. \code{qstnr_lbl_col_sym()} is intended to ease language-agnostic questionnaire data
programming.
}
\examples{
fokus:::qstnr_lbl_col_sym(lang = "de")

fokus:::qstnr_lbl_col_sym(lang = "en")
}
\concept{qstnr_predicate}
\keyword{internal}
