% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{qstnr_response_option_codes}
\alias{qstnr_response_option_codes}
\title{Get response option codes}
\usage{
qstnr_response_option_codes(types = all_response_option_types)
}
\arguments{
\item{types}{Response option types. One or more of
\itemize{
\item \code{"abstain"}
\item \code{"custom"}
\item \code{"dunno"}
\item \code{"no"}
\item \code{"none"}
\item \code{"other"}
\item \code{"refusal"}
\item \code{"unspecific"}
\item \code{"yes"}
}}
}
\value{
An integer vector.
}
\description{
Extracts response option codes of the specified type(s) from the \link[=raw_qstnr]{raw FOKUS questionnaire data}.
}
\details{
Note that only codes of recurring response options are returned which are defined under the \code{response_options} top-level key in the file
\code{data-raw/questionnaire/questionnaire.toml}.
}
\examples{
fokus:::qstnr_response_option_codes(types = "abstain")

fokus:::qstnr_response_option_codes(types = c("dunno",
                                              "custom",
                                              "abstain"))
}
\seealso{
Other questionnaire-internal functions:
\code{\link{qstnr_item_val}()},
\code{\link{qstnr_parties}()}
}
\concept{qstnr_internal}
\keyword{internal}
