% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{qx}
\alias{qx}
\title{Questionnaire data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
qx
}
\description{
A tibble containing the data of all FOKUS questionnaires.
}
\details{
\code{qx} was generated based on the following steps:
\enumerate{
\item \code{\link[=gen_q_tibble]{gen_q_tibble()}} was run for all valid combinations of \code{canton} and \code{ballot_date}, some validation checks were performed, and the results were merged into
a single tibble.
\item List columns were \link[=expand_q_tibble]{expanded} to \href{https://en.wikipedia.org/wiki/Wide_and_narrow_data}{long format}.
\item Columns \href{https://rpkg.dev/fokus/articles/raw_q_schema.html#supported-keys}{\code{question_intro_i} and \code{question_intro_j}} were merged into the single column
\code{question_intro}.
\item Column \href{https://rpkg.dev/fokus/articles/raw_q_schema.html#supported-keys}{\code{question_full}} was complemented, i.e. made to fall back on \code{question} if \code{NA}.
\item Columns \href{https://rpkg.dev/fokus/articles/raw_q_schema.html#supported-keys}{\code{question_common}} and
\href{https://rpkg.dev/fokus/articles/raw_q_schema.html#supported-keys}{\code{variable_label_common}} were complemented, i.e. made to fall back on \code{question_full}
and \code{variable_label} respectively if \code{NA}.
\item Markdown formatting was \link[pal:strip_md]{stripped} from all character columns.
}
}
\seealso{
Other questionnaire and survey functions and datasets:
\code{\link{elections}},
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_q_all}()},
\code{\link{export_qr_codes}()},
\code{\link{export_q}()},
\code{\link{proposals}}
}
\concept{q_survey}
\keyword{datasets}
