% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{qx_suppl}
\alias{qx_suppl}
\title{Raw supplemental date-specific FOKUS questionnaire data}
\format{
A \link[xfun:strict_list]{strict list}.
}
\usage{
qx_suppl
}
\description{
A structured list of raw supplemental date-specific questionnaire data of the FOKUS surveys.
}
\examples{
fokus::qx_suppl[["2018-09-23"]]$mode
}
\seealso{
\code{\link{q}} \code{\link{q_suppl}} \code{\link{q_suppl_lvl}} \code{\link{q_suppl_lvl_canton}} \code{\link{q_suppl_proposal}}
\code{\link{q_suppl_election}} \code{\link{q_suppl_skill_questions}} \code{\link{q_suppl_skill_question}}
}
\keyword{datasets}
