% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{qx_supplemental}
\alias{qx_supplemental}
\title{Raw supplemental date-specific FOKUS questionnaire data}
\format{
A \link[xfun:strict_list]{strict list}.
}
\usage{
qx_supplemental
}
\description{
A structured list of raw supplemental date-specific questionnaire data of the FOKUS surveys.
}
\seealso{
\code{\link{q}} \code{\link[=gen_q]{gen_q()}} \code{\link[=q_tibble]{q_tibble()}} \code{\link[=q_md]{q_md()}} \code{\link[=pick_right]{pick_right()}}
}
\keyword{datasets}
