% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_q_suppl_arguments}
\alias{raw_q_suppl_arguments}
\title{Get a proposal's raw argument supplemental questionnaire data}
\usage{
raw_q_suppl_arguments(
  ballot_date = all_ballot_dates,
  lvl = all_lvls,
  canton = all_cantons,
  proposal_nr = 1L
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{proposal_nr}{Proposal number. A positive integer scalar.}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of argument data from the \link[=raw_q_suppl_proposal]{proposal-specific raw supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus:::proposal_main_motives(ballot_date = "2018-11-25",
                              lvl = "cantonal",
                              canton = "aargau",
                              proposal_nr = 2)
}
\seealso{
Raw questionnaire data \code{\link{raw_q}} \code{\link{raw_qx_suppl}}

Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_q_suppl_election_name}()},
\code{\link{raw_q_suppl_elections}()},
\code{\link{raw_q_suppl_election}()},
\code{\link{raw_q_suppl_lvl_canton}()},
\code{\link{raw_q_suppl_lvl}()},
\code{\link{raw_q_suppl_main_motives}()},
\code{\link{raw_q_suppl_mode}()},
\code{\link{raw_q_suppl_proposal_name}()},
\code{\link{raw_q_suppl_proposal}()},
\code{\link{raw_q_suppl_skill_questions}()},
\code{\link{raw_q_suppl_skill_question}()},
\code{\link{raw_q_suppl}()}
}
\concept{q_raw}
\keyword{internal}
