% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_q_suppl_election}
\alias{raw_q_suppl_election}
\title{Get an election's raw supplemental questionnaire data}
\usage{
raw_q_suppl_election(
  ballot_date = all_ballot_dates,
  lvl = all_lvls,
  canton = cantons(ballot_date),
  prcd = all_prcds,
  election_nr = 1L
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{prcd}{Election procedure. One of
\itemize{
\item \code{"proportional"}
\item \code{"majoritarian"}
}}

\item{election_nr}{Election number. A positive integer scalar (in almost all cases \code{1L}).}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of an election's data from the \link[=raw_q_suppl_lvl_canton]{canton's political-level-specific raw supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus:::raw_q_suppl_election(ballot_date = "2019-10-20",
                             lvl = "cantonal",
                             canton = "aargau",
                             prcd = "majoritarian") \%$\%
  n_seats
}
\seealso{
Raw questionnaire data \code{\link{raw_q}} \code{\link{raw_qx_suppl}}

Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_q_suppl_arguments}()},
\code{\link{raw_q_suppl_election_name}()},
\code{\link{raw_q_suppl_elections}()},
\code{\link{raw_q_suppl_lvl_canton}()},
\code{\link{raw_q_suppl_lvl}()},
\code{\link{raw_q_suppl_main_motives}()},
\code{\link{raw_q_suppl_mode}()},
\code{\link{raw_q_suppl_proposal_name}()},
\code{\link{raw_q_suppl_proposal}()},
\code{\link{raw_q_suppl_skill_questions}()},
\code{\link{raw_q_suppl_skill_question}()},
\code{\link{raw_q_suppl}()}
}
\concept{q_raw}
\keyword{internal}
