% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_q_suppl_lvl}
\alias{raw_q_suppl_lvl}
\title{Get a political level's raw supplemental date-specific FOKUS questionnaire data}
\usage{
raw_q_suppl_lvl(ballot_date = ballot_dates, lvl = c("cantonal", "federal"))
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of the \link[=raw_q_suppl]{raw supplemental date-specific FOKUS questionnaire data} for the specified ballot date and political level.
}
\examples{
fokus:::raw_q_suppl_lvl(ballot_date = "2018-09-23",
                        lvl = "cantonal") |>
  names()
}
\seealso{
Raw questionnaire data \code{\link{raw_q}} \code{\link{raw_qx_suppl}}

Other raw questionnaire data functions
\code{\link{raw_pick_right}()},
\code{\link{raw_q_suppl_arguments}()},
\code{\link{raw_q_suppl_election_name}()},
\code{\link{raw_q_suppl_elections}()},
\code{\link{raw_q_suppl_election}()},
\code{\link{raw_q_suppl_lvl_canton}()},
\code{\link{raw_q_suppl_main_motives}()},
\code{\link{raw_q_suppl_mode}()},
\code{\link{raw_q_suppl_proposal_name}()},
\code{\link{raw_q_suppl_proposal}()},
\code{\link{raw_q_suppl_skill_questions}()},
\code{\link{raw_q_suppl_skill_question}()},
\code{\link{raw_q_suppl}()}
}
\concept{q_raw}
\keyword{internal}
