% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_q_suppl_mode}
\alias{raw_q_suppl_mode}
\title{Get raw supplemental survey mode questionnaire data}
\usage{
raw_q_suppl_mode(ballot_date = all_ballot_dates, canton = cantons(ballot_date))
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of survey mode data from the \link[=raw_q_suppl]{raw supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus:::raw_q_suppl_mode(ballot_date = "2018-09-23",
                         canton = "aargau")
}
\seealso{
Raw questionnaire data \code{\link{raw_q}} \code{\link{raw_qx_suppl}}

Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_q_suppl_arguments}()},
\code{\link{raw_q_suppl_election_name}()},
\code{\link{raw_q_suppl_elections}()},
\code{\link{raw_q_suppl_election}()},
\code{\link{raw_q_suppl_lvl_canton}()},
\code{\link{raw_q_suppl_lvl}()},
\code{\link{raw_q_suppl_main_motives}()},
\code{\link{raw_q_suppl_proposal_name}()},
\code{\link{raw_q_suppl_proposal}()},
\code{\link{raw_q_suppl_skill_questions}()},
\code{\link{raw_q_suppl_skill_question}()},
\code{\link{raw_q_suppl}()}
}
\concept{q_raw}
\keyword{internal}
