% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_q_suppl_proposal}
\alias{raw_q_suppl_proposal}
\title{Get raw proposal supplemental questionnaire data}
\usage{
raw_q_suppl_proposal(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  proposal_nr
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{proposal_nr}{Proposal number. A positive integer scalar or \code{NULL}. If \code{NULL}, it is considered to be a non-proposal-specific skill question
(the case at elections).}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of a proposal's data from the \link[=raw_q_suppl]{supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus:::raw_q_suppl_proposal(ballot_date = "2018-09-23",
                             lvl = "cantonal",
                             canton = "aargau",
                             proposal_nr = 1) \%$\%
  name
}
\seealso{
Raw questionnaire data \code{\link{raw_q}} \code{\link{raw_qx_suppl}}

Other raw questionnaire data functions
\code{\link{raw_pick_right}()},
\code{\link{raw_q_suppl_arguments}()},
\code{\link{raw_q_suppl_election_name}()},
\code{\link{raw_q_suppl_elections}()},
\code{\link{raw_q_suppl_election}()},
\code{\link{raw_q_suppl_lvl_canton}()},
\code{\link{raw_q_suppl_lvl}()},
\code{\link{raw_q_suppl_main_motives}()},
\code{\link{raw_q_suppl_proposal_name}()},
\code{\link{raw_q_suppl_skill_questions}()},
\code{\link{raw_q_suppl_skill_question}()},
\code{\link{raw_q_suppl}()}
}
\concept{q_raw}
\keyword{internal}
