% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_qstnr_suppl_elections}
\alias{raw_qstnr_suppl_elections}
\title{Get raw supplemental election questionnaire data}
\usage{
raw_qstnr_suppl_elections(
  ballot_date = all_ballot_dates,
  lvl = all_lvls,
  canton = all_cantons
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of election data from the \link[=raw_qstnr_suppl_lvl_canton]{canton's political-level-specific raw supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus:::raw_qstnr_suppl_elections(ballot_date = "2019-10-20",
                                  lvl = "cantonal",
                                  canton = "aargau") \%$\%
  skill_questions_source
}
\seealso{
Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_qstnr_suppl_arguments}()},
\code{\link{raw_qstnr_suppl_argument}()},
\code{\link{raw_qstnr_suppl_election_name}()},
\code{\link{raw_qstnr_suppl_election}()},
\code{\link{raw_qstnr_suppl_lvl_canton}()},
\code{\link{raw_qstnr_suppl_lvl}()},
\code{\link{raw_qstnr_suppl_main_motives}()},
\code{\link{raw_qstnr_suppl_mode}()},
\code{\link{raw_qstnr_suppl_proposal_name}()},
\code{\link{raw_qstnr_suppl_proposal}()},
\code{\link{raw_qstnr_suppl_skill_questions}()},
\code{\link{raw_qstnr_suppl_skill_question}()},
\code{\link{raw_qstnr_suppl}()},
\code{\link{raw_qstnrs_suppl}},
\code{\link{raw_qstnr}}
}
\concept{qstnr_raw}
\keyword{internal}
