% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_qstnr_suppl_skill_question}
\alias{raw_qstnr_suppl_skill_question}
\title{Get a single skill question's raw supplemental questionnaire data}
\usage{
raw_qstnr_suppl_skill_question(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = all_lvls,
  canton = pal::pkg_config_val(key = "canton", pkg = this_pkg),
  proposal_nr = NULL,
  skill_question_nr = 1L
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{proposal_nr}{Proposal number. A positive integerish scalar or \code{NULL}. If \code{NULL}, it is considered to be a non-proposal-specific skill question
(the case at elections).}

\item{skill_question_nr}{Skill question number. A positive integerish scalar.}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of a single skill question's data from the \link[=raw_qstnr_suppl]{raw supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus:::raw_qstnr_suppl_skill_question(ballot_date = "2018-09-23",
                                       lvl = "cantonal",
                                       canton = "aargau",
                                       proposal_nr = 1,
                                       skill_question_nr = 1)
}
\seealso{
Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_qstnr}},
\code{\link{raw_qstnr_suppl}()},
\code{\link{raw_qstnr_suppl_argument}()},
\code{\link{raw_qstnr_suppl_arguments}()},
\code{\link{raw_qstnr_suppl_election}()},
\code{\link{raw_qstnr_suppl_election_name}()},
\code{\link{raw_qstnr_suppl_elections}()},
\code{\link{raw_qstnr_suppl_lvl}()},
\code{\link{raw_qstnr_suppl_lvl_canton}()},
\code{\link{raw_qstnr_suppl_main_motives}()},
\code{\link{raw_qstnr_suppl_mode}()},
\code{\link{raw_qstnr_suppl_proposal}()},
\code{\link{raw_qstnr_suppl_proposal_name}()},
\code{\link{raw_qstnr_suppl_proposal_safe}()},
\code{\link{raw_qstnr_suppl_skill_questions}()},
\code{\link{raw_qstnrs_suppl}}
}
\concept{qstnr_raw}
\keyword{internal}
