% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{raw_qstnr_suppl_skill_questions}
\alias{raw_qstnr_suppl_skill_questions}
\title{Get raw supplemental skill question questionnaire data}
\usage{
raw_qstnr_suppl_skill_questions(
  ballot_date = all_ballot_dates,
  lvl = all_lvls,
  canton = all_cantons,
  proposal_nr = NULL
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{proposal_nr}{Proposal number. A positive integerish scalar or \code{NULL}. If \code{NULL}, non-proposal-specific skill question data is returned.}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Returns a structured list of skill question data from the \link[=raw_qstnr_suppl]{raw supplemental date-specific FOKUS questionnaire data}.
}
\examples{
fokus:::raw_qstnr_suppl_skill_questions(ballot_date = "2018-09-23",
                                        lvl = "cantonal",
                                        canton = "aargau",
                                        proposal_nr = 1) |>
  purrr::map_depth(1L, "de") |>
  purrr::list_c(ptype = character())
}
\seealso{
Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_qstnr_suppl_arguments}()},
\code{\link{raw_qstnr_suppl_argument}()},
\code{\link{raw_qstnr_suppl_election_name}()},
\code{\link{raw_qstnr_suppl_elections}()},
\code{\link{raw_qstnr_suppl_election}()},
\code{\link{raw_qstnr_suppl_lvl_canton}()},
\code{\link{raw_qstnr_suppl_lvl}()},
\code{\link{raw_qstnr_suppl_main_motives}()},
\code{\link{raw_qstnr_suppl_mode}()},
\code{\link{raw_qstnr_suppl_proposal_name}()},
\code{\link{raw_qstnr_suppl_proposal}()},
\code{\link{raw_qstnr_suppl_skill_question}()},
\code{\link{raw_qstnr_suppl}()},
\code{\link{raw_qstnrs_suppl}},
\code{\link{raw_qstnr}}
}
\concept{qstnr_raw}
\keyword{internal}
