% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{raw_qstnrs_suppl}
\alias{raw_qstnrs_suppl}
\title{Raw supplemental date-specific FOKUS questionnaire data}
\format{
A \link[xfun:strict_list]{strict list}.
}
\description{
A structured list of raw supplemental date-specific questionnaire data of the FOKUS surveys.
}
\examples{
fokus:::raw_qstnrs_suppl[["2018-09-23"]]$mode
}
\seealso{
Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_qstnr_suppl_arguments}()},
\code{\link{raw_qstnr_suppl_argument}()},
\code{\link{raw_qstnr_suppl_election_name}()},
\code{\link{raw_qstnr_suppl_elections}()},
\code{\link{raw_qstnr_suppl_election}()},
\code{\link{raw_qstnr_suppl_lvl_canton}()},
\code{\link{raw_qstnr_suppl_lvl}()},
\code{\link{raw_qstnr_suppl_main_motives}()},
\code{\link{raw_qstnr_suppl_mode}()},
\code{\link{raw_qstnr_suppl_proposal_name}()},
\code{\link{raw_qstnr_suppl_proposal}()},
\code{\link{raw_qstnr_suppl_skill_questions}()},
\code{\link{raw_qstnr_suppl_skill_question}()},
\code{\link{raw_qstnr_suppl}()},
\code{\link{raw_qstnr}}
}
\concept{qstnr_raw}
\keyword{internal}
