% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_easyvote_municipalities}
\alias{read_easyvote_municipalities}
\title{Read in easyvote municipality data}
\usage{
read_easyvote_municipalities(ballot_date, canton)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reads in the latest dataset of easyvote municipality information provided to us not earlier than 90 days before and up until 20 days after the \code{ballot_date}.
}
\details{
If both columns \code{min_age} and \code{max_age} are \code{NA} in the data returned, it means that
\itemize{
\item the municipality did not provide easyvote with specific information on the target age range, and
\item the municipality has subscribed to parcel mailing (instead of direct delivery to households) and delivers the brochures itself -- very likely to young
adults between 18--25 years.
}
}
\examples{
# private FOKUS directory needs to be accessible for this function to work
try(
  fokus:::read_easyvote_municipalities(ballot_date = "2020-09-27",
                                       canton = "aargau")
)
}
\keyword{internal}
