% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_online_participation_codes}
\alias{read_online_participation_codes}
\title{Read in online participation codes}
\usage{
read_online_participation_codes(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A character vector.
}
\description{
Reads in the online participation codes externally generated by the survey institute that are necessary to \link[=export_qr_codes]{generate QR codes with personalized survey URLs}.
}
\examples{
# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_online_participation_codes(ballot_date = "2018-11-25",
                                         canton = "aargau")
)
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_private_file}()},
\code{\link{read_voting_register_data_extra}()},
\code{\link{read_voting_register_ids}()}
}
\concept{data_import}
