% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_private_file}
\alias{read_private_file}
\title{Read in file from private FOKUS repository}
\usage{
read_private_file(
  path,
  as_chr = FALSE,
  use_cache = TRUE,
  auth_token = pal::pkg_config_val("token_repo_private")
)
}
\arguments{
\item{path}{File path relative to the repository root.}

\item{as_chr}{Whether to return the file content as a character scalar, which is only recommended for text files. If \code{FALSE}, the file is returned as a
\link[=raw]{raw vector} instead.}

\item{use_cache}{Whether or not to return cached results if possible. Caching is done based on file content hashing, so the file is only newly fetched if it
actually changed since the last download. If \code{FALSE}, the file is always newly fetched.}

\item{auth_token}{\href{https://gitlab.com/-/user_settings/personal_access_tokens}{Personal access token of a gitlab.com account} with access to the \href{https://gitlab.com/c2d-zda/private/fokus_private/}{private FOKUS repository}.}
}
\value{
The file content, as character scalar if \code{as_chr}, otherwise as a \link[=raw]{raw vector}.
}
\description{
Downloads a file from the \href{https://gitlab.com/c2d-zda/private/fokus_private/}{private FOKUS repository} and returns it as a raw vector or a character scalar, depending on \code{as_chr}.
}
\details{
Files are downloaded via \href{https://docs.gitlab.com/ee/api/rest/}{GitLab's RESTful API (v4)}. If \code{use_cache = TRUE} (the default), a downloaded file is cached
on disk in this package's \link[pkgpins:board]{user-cache pins board} and only newly fetched from the private FOKUS repository GitLab remote if is has changed
since being downloaded the last time. Caching saves a bit of time and (potentially) a lot of bandwidth.
}
\examples{
# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_private_file("raw/survey_data_2018-09-23_aargau.xlsx") |>
    length()
)
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_online_participation_codes}()},
\code{\link{read_survey_data}()},
\code{\link{read_voting_register_data_extra}()},
\code{\link{read_voting_register_ids}()},
\code{\link{tidy_cols}()}

Other private FOKUS repository functions:
\code{\link{print_private_repo_structure}()},
\code{\link{url_repo_private}()},
\code{\link{write_private_file}()}
}
\concept{data_import}
\concept{private}
