% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{read_q}
\alias{read_q}
\title{Read in raw questionnaire data}
\usage{
read_q()
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Reads in file \code{questionnaire/questionnaire.toml} and returns it as a structured list.
}
\seealso{
Questionnaire-related functions
\code{\link{gen_q}()},
\code{\link{pick_right}()},
\code{\link{q_md}()},
\code{\link{q_tibble}()},
\code{\link{read_q_supplemental}()},
\code{\link{read_toml}()}
}
\concept{questionnaire}
