% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_q_supplemental}
\alias{read_q_supplemental}
\title{Read in raw supplemental date-specific questionnaire data}
\usage{
read_q_supplemental(ballot_date = as.character(ballot_dates))
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"} )`
}}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Reads in file \code{questionnaire/YYYY-MM-DD.toml} where \code{YYYY-MM-DD} corresponds to the \code{ballot_date} argument and returns it as a structured list.
}
\seealso{
Questionnaire-related functions
\code{\link{gen_q}()},
\code{\link{pick_right}()},
\code{\link{q_md}()},
\code{\link{q_tibble}()},
\code{\link{read_q}()}
}
\concept{questionnaire}
