% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_survey_data}
\alias{read_survey_data}
\title{Read in generated survey data}
\usage{
read_survey_data(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  lang = all_langs
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reads in the FOKUS survey data generated by \code{\link[=export_survey_data]{export_survey_data()}} for the specified ballot date in the specified canton and the specified language.
}
\examples{
# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_survey_data(ballot_date = "2023-06-18",
                          canton = "aargau",
                          lang = "en")
)
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_online_participation_codes}()},
\code{\link{read_private_file}()},
\code{\link{read_voting_register_data_extra}()},
\code{\link{read_voting_register_ids}()}
}
\concept{data_import}
