% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_toml}
\alias{read_toml}
\title{Read in and parse a TOML file as a strict list}
\usage{
read_toml(path)
}
\arguments{
\item{path}{Path to a TOML file. A character scalar.}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Reads in a file in \href{https://toml.io/}{Tom's Obvious Minimal Language} format and returns its content as a (nested) \link[xfun:strict_list]{strict list}.
}
\details{
The file is parsed using \code{\link[RcppTOML:parseTOML]{RcppTOML::parseTOML(escape = FALSE)}}.
}
