% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_voting_register_data_extra}
\alias{read_voting_register_data_extra}
\title{Read in extra voting register data}
\usage{
read_voting_register_data_extra(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  use_cache = TRUE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{use_cache}{Whether or not to return cached results if possible. Caching is done based on file content hashing, so the file is only newly fetched if it
actually changed since the last download. If \code{FALSE}, the file is always newly fetched.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reads in the raw extra voting register data provided by the statistical office, performs various integrity checks and returns it in tidy shape.
}
\examples{
# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_voting_register_data_extra(ballot_date = "2019-10-20",
                                         canton = "aargau") |>
    colnames()
)
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_online_participation_codes}()},
\code{\link{read_private_file}()},
\code{\link{read_survey_data}()},
\code{\link{read_voting_register_ids}()}
}
\concept{data_import}
