% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_voting_register_ids}
\alias{read_voting_register_ids}
\title{Read in voting register identifiers}
\usage{
read_voting_register_ids(
  ballot_date = pal::pkg_config_val("ballot_date"),
  canton = cantons(ballot_date),
  use_cache = TRUE,
  auth_token = pal::pkg_config_val("token_repo_private")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{use_cache}{Whether or not to return cached results if possible. Caching is done based on file content hashing, so the file is only newly fetched if it
actually changed since the last download. If \code{FALSE}, the file is always newly fetched.}

\item{auth_token}{\href{https://gitlab.com/-/user_settings/personal_access_tokens}{Personal access token of a gitlab.com account} with access to the \href{https://gitlab.com/c2d-zda/private/fokus_private/}{private FOKUS repository}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the column \code{id_voting_register}.
}
\description{
Reads in the voting register identifiers of the population that was invited to participate in the FOKUS survey at the specified ballot date in the specified
canton.
}
\details{
Note that this data is not available for all FOKUS surveys.
}
\examples{
# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_voting_register_ids(ballot_date = "2019-10-20",
                                  canton = "aargau") |>
    nrow()
)
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_online_participation_codes}()},
\code{\link{read_private_file}()},
\code{\link{read_survey_data}()},
\code{\link{read_voting_register_data_extra}()},
\code{\link{tidy_cols}()}
}
\concept{data_import}
