% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{requires_candidate_registration}
\alias{requires_candidate_registration}
\title{Determine whether majoritarian election requires candidate registration}
\usage{
requires_candidate_registration(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  election_nr = 1L
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{Election number. A positive integer scalar (in almost all cases \code{1L}).}
}
\value{
A logical scalar.
}
\description{
Determines whether or not candidates must be officially registered prior to the specified majority election.
}
\details{
The absence of a candidate registration requirement usually means that every eligible citizen can be elected, i.e. receive valid votes.
}
\examples{
fokus::requires_candidate_registration(ballot_date = "2019-10-20",
                                       lvl = "federal",
                                       canton = "aargau")
}
\seealso{
Other election predicate functions:
\code{\link{election_candidate_string}()},
\code{\link{election_candidates}()},
\code{\link{election_names_combined}()},
\code{\link{election_name}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()}
}
\concept{predicate_election}
