% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{response_option_types}
\alias{response_option_types}
\title{List available response option types}
\format{
A character vector.
}
\usage{
response_option_types
}
\description{
A vector of all response option types defined in the \link[=raw_q]{raw FOKUS questionnaire data}.
}
\seealso{
\code{\link{ballot_dates}} \code{\link{cantons}} \code{\link{proposal_types}} \code{\link{q_item_keys}}
}
\keyword{datasets}
