% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{response_options}
\alias{response_options}
\title{Get response options}
\usage{
response_options(
  type = all_response_option_types,
  lang = pal::pkg_config_val("lang"),
  subtypes = NULL
)
}
\arguments{
\item{type}{Response option type. One of
\itemize{
\item \code{"abstain"}
\item \code{"custom"}
\item \code{"dunno"}
\item \code{"no"}
\item \code{"none"}
\item \code{"other"}
\item \code{"refusal"}
\item \code{"unspecific"}
\item \code{"yes"}
}}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}

\item{subtypes}{Hierarchical response option subtypes as a character vector, or \code{NULL}. If \code{NULL}, all subtypes are returned.}
}
\value{
A character vector.
}
\description{
Returns response options of the specified type and optionally subtype(s) from the \link[=raw_qstnr]{raw FOKUS questionnaire data}.
}
\details{
Note that only recurring response options are returned which are defined under the \code{response_options} top-level key in the file
\code{data-raw/questionnaire/questionnaire.toml}.
}
\examples{
fokus:::response_options(type = "abstain",
                         lang = "de")

fokus:::response_options(type = "abstain",
                         lang = "de",
                         subtypes = "election")

fokus:::response_options(type = "abstain",
                         lang = "de",
                         subtypes = c("election", "proportional"))
}
\seealso{
Other predicate functions:
\code{\link{ballot_title}()},
\code{\link{political_issues}()},
\code{\link{postal_dispatch_way}()}
}
\concept{predicate_other}
