% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{shorten_v_names}
\alias{shorten_v_names}
\title{Shorten variable names to a maximum length of 32 characters}
\usage{
shorten_v_names(v_names, reverse = FALSE, max_n_char = 32L)
}
\arguments{
\item{v_names}{A character vector of variable names.}

\item{reverse}{Whether to apply the inversion of the shortening logic, i.e. to restore original/unshortened variable names.}

\item{max_n_char}{Maximum allowed number of characters. Either \code{NULL} to skip the check or otherwise an integer scalar in which case it is ensured that the
maximum resulting variable name length doesn't exceed it. Doesn't have any influence on the applied shortening logic.}
}
\value{
A character vector of the same length as \code{v_names}.
}
\description{
Shorten variable names to a maximum length of 32 characters
}
\seealso{
\code{\link[=shorten_colnames]{shorten_colnames()}}, \code{\link[=restore_colnames]{restore_colnames()}}

Other variable-related functions:
\code{\link{is_skill_question_v}()},
\code{\link{v_lbl}()},
\code{\link{v_lvls}()},
\code{\link{v_proposal_nr}()},
\code{\link{v_skill_question_nr}()}
}
\concept{variable}
