% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{shorten_var_names}
\alias{shorten_var_names}
\title{Shorten variable names to a maximum length of 32 characters}
\usage{
shorten_var_names(var_names, reverse = FALSE, max_n_char = 32L)
}
\arguments{
\item{var_names}{A character vector of variable names.}

\item{reverse}{Whether to apply the inversion of the shortening logic, i.e. to restore original/unshortened variable names.}

\item{max_n_char}{Maximum allowed number of characters. Either \code{NULL} to skip the check or otherwise an integerish scalar in which case it is ensured that
the maximum resulting variable name length doesn't exceed it. Doesn't have any influence on the applied shortening logic.}
}
\value{
A character vector of the same length as \code{var_names}.
}
\description{
Shorten variable names to a maximum length of 32 characters
}
\seealso{
Other variable name shortening functions:
\code{\link{restore_colnames}()},
\code{\link{shorten_colnames}()}
}
\concept{var_name_shortening}
