% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{skill_question}
\alias{skill_question}
\title{Get skill question}
\usage{
skill_question(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  proposal_nr = NULL,
  skill_question_nr = 1L,
  lang = pal::pkg_config_val("lang")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integerish scalar or \code{NULL}. If \code{NULL}, it is considered to be a non-proposal-specific skill question
(the case at elections).}

\item{skill_question_nr}{Skill question number. A positive integerish scalar.}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A character scalar.
}
\description{
Returns the skill question text in the specified language. Note that by default (\code{proposal_nr = NULL}), only non-proposal-specific skill questions are
returned.
}
\examples{
fokus::skill_question(ballot_date = "2018-09-23",
                      lvl = "cantonal",
                      canton = "aargau",
                      proposal_nr = 1,
                      skill_question_nr = 2,
                      lang = "en")
}
\seealso{
Other skill question predicate functions:
\code{\link{n_skill_questions}()},
\code{\link{skill_question_answer_nr}()},
\code{\link{skill_question_nrs}()},
\code{\link{skill_question_proposal_nrs}()},
\code{\link{skill_question_response_options}()}
}
\concept{predicate_skill_question}
