% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{skill_question_nrs}
\alias{skill_question_nrs}
\title{Get skill question numbers}
\usage{
skill_question_nrs(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  proposal_nr = NULL
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integerish scalar or \code{NULL}. If \code{NULL}, the numbers of non-proposal-specific skill questions are returned.}
}
\value{
An integer vector.
}
\description{
Determines the skill questions numbers at the specified ballot date on the specified political level. Note that by default (\code{proposal_nr = NULL}),
non-proposal-specific skill question numbers are returned.
}
\examples{
fokus::skill_question_nrs(ballot_date = "2018-09-23",
                          lvl = "cantonal",
                          canton = "aargau",
                          proposal_nr = 1)

# note that by default, non-proposal-specific skill question numbers are returned
fokus::skill_question_nrs(ballot_date = "2018-09-23",
                          lvl = "cantonal",
                          canton = "aargau")

fokus::skill_question_nrs(ballot_date = "2019-10-20",
                          lvl = "cantonal",
                          canton = "aargau")
}
\seealso{
Other skill question predicate functions:
\code{\link{n_skill_questions}()},
\code{\link{skill_question_answer_nr}()},
\code{\link{skill_question_proposal_nrs}()},
\code{\link{skill_question_response_options}()},
\code{\link{skill_question}()}
}
\concept{predicate_skill_question}
