% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{skill_question_proposal_nrs}
\alias{skill_question_proposal_nrs}
\title{Get referendum proposal numbers with skill questions}
\usage{
skill_question_proposal_nrs(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "referendum"),
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}
}
\value{
An integer vector.
}
\description{
Determines the referendum proposal numbers covered by the FOKUS survey for the specified canton at the specified ballot date on the specified political
level that have at least one skill question.
}
\examples{
fokus::skill_question_proposal_nrs(ballot_date = "2018-09-23",
                                   lvl = "cantonal",
                                   canton = "aargau")
}
\seealso{
Other skill question predicate functions:
\code{\link{n_skill_questions}()},
\code{\link{skill_question}()},
\code{\link{skill_question_answer_nr}()},
\code{\link{skill_question_nrs}()},
\code{\link{skill_question_response_options}()}
}
\concept{predicate_skill_question}
