% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{tidy_cols}
\alias{tidy_cols}
\title{Tidy columns}
\usage{
tidy_cols(data, default_type = readr::col_character())
}
\arguments{
\item{data}{Data frame whose columns are to be converted.}

\item{default_type}{\link[readr:parse_atomic]{readr column type} to which columns are to be converted for which no explicit type is predefined.}
}
\value{
\code{data}
}
\description{
Renames a data frame's columns and converts them to their proper types. Aimed at datasets published by cantonal statistical offices.
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_online_participation_codes}()},
\code{\link{read_private_file}()},
\code{\link{read_survey_data}()},
\code{\link{read_voting_register_data_extra}()},
\code{\link{read_voting_register_ids}()}
}
\concept{data_import}
