% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{v_lbl}
\alias{v_lbl}
\title{Get variable label}
\usage{
v_lbl(v_name, ballot_date = NULL, canton = NULL)
}
\arguments{
\item{v_name}{Variable name. A character scalar.}

\item{ballot_date}{\code{NULL} or a valid FOKUS-covered cantonal ballot date, i.e. one of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{\code{NULL} or a valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}
}
\value{
A character scalar.
}
\description{
Extracts a variable's (common) label from the \link[=qx]{questionnaire data}.
}
\details{
If no \code{ballot_date} and \code{canton} are specified, \href{https://rpkg.dev/fokus/articles/raw_q_schema.html#supported-keys}{\code{variable_label_common}} is returned,
otherwise \href{https://rpkg.dev/fokus/articles/raw_q_schema.html#supported-keys}{\code{variable_label}}. Note that \code{ballot_date} and \code{canton} either must both be
\code{NULL} or set to a valid canton name and ballot date respectively.
}
\examples{
# if `ballot_date` and `canton` are not supplied, `variable_label_common` is returned:
fokus::v_lbl("participation_federal")

# otherwise `variable_label`:
fokus::v_lbl(v_name = "participation_federal",
             ballot_date = "2018-09-23",
             canton = "aargau")
            
fokus::v_lbl(v_name = "participation_federal",
             ballot_date = "2019-10-20",
             canton = "aargau")
}
