% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{v_lvls}
\alias{v_lvls}
\title{Determine variable's political level(s)}
\usage{
v_lvls(v_name)
}
\arguments{
\item{v_name}{Variable name. A character scalar.}
}
\value{
A character vector of political levels, of length 0 if no specific levels could be determined.
}
\description{
Determines a variable's political level(s).
}
\details{
Note that the political levels are determined by simply parsing \code{v_name}.
}
\examples{
fokus::v_lvls("skill_question_1_cantonal_proposal_1")
}
\seealso{
Other variable-related functions:
\code{\link{is_skill_question_v}()},
\code{\link{shorten_v_names}()},
\code{\link{v_proposal_nr}()},
\code{\link{v_skill_question_nr}()}
}
\concept{variable}
