% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{v_proposal_nr}
\alias{v_proposal_nr}
\title{Determine variable's proposal number}
\usage{
v_proposal_nr(v_names)
}
\arguments{
\item{v_names}{A character vector of FOKUS variable names.}
}
\value{
An integer vector of the same length as \code{v_names}.
}
\description{
Determines the proposal number each variable corresponds to. In case no proposal number could be determined for a variable, \code{NA_character} is returned.
}
\details{
Note that the determination is performed by simply parsing \code{v_names}.
}
\seealso{
Other variable-related functions:
\code{\link{is_skill_question_v}()},
\code{\link{shorten_v_names}()},
\code{\link{v_lvls}()},
\code{\link{v_skill_question_nr}()}
}
\concept{variable}
