% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{v_val_lbls}
\alias{v_val_lbls}
\title{Get pre-defined variable value labels}
\usage{
v_val_lbls(type = c("non_material_participants"), lang = c("de", "en"))
}
\arguments{
\item{type}{Variable value type.}

\item{lang}{Language the labels should be returned in.}
}
\value{
A character vector.
}
\description{
Get pre-defined variable value labels
}
\seealso{
FOKUS-variable-related functions
\code{\link{is_skill_question}()},
\code{\link{shorten_v_names}()}
}
\concept{variable}
