% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_election_nr}
\alias{var_election_nr}
\title{Determine variable's election number}
\usage{
var_election_nr(var_names)
}
\arguments{
\item{var_names}{A character vector of FOKUS variable names.}
}
\value{
An integer vector of the same length as \code{var_names}.
}
\description{
Determines the election number each variable corresponds to. In case no election number could be determined for a variable, \code{NA_character} is returned.
}
\details{
Note that the determination is performed by simply parsing \code{var_names}.
}
\examples{
fokus::qstnrs |>
  dplyr::mutate(election_nr = fokus::var_election_nr(variable_name)) |>
  dplyr::filter(!is.na(election_nr)) |>
  dplyr::select(variable_name, election_nr) |>
  unique()
}
\seealso{
Other variable-related functions:
\code{\link{add_vars_to_combos}()},
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_lvls}()},
\code{\link{var_prcd}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_title}()},
\code{\link{var_val_set}()}
}
\concept{vars}
