% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_lbl}
\alias{var_lbl}
\title{Get variable label}
\usage{
var_lbl(var_name, ballot_date = NULL, canton = NULL)
}
\arguments{
\item{var_name}{Variable name. A character scalar.}

\item{ballot_date}{\code{NULL} or a FOKUS-covered ballot date, i.e. one of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{canton}{\code{NULL} or a canton name, i.e. one of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A character scalar.
}
\description{
Extracts a variable's (common) label from the \link[=qstnrs]{questionnaire data}.
}
\details{
If no \code{ballot_date} and \code{canton} are specified, \href{https://fokus.rpkg.dev/articles/raw_qstnr_schema.html#supported-keys}{\code{variable_label_common}} is returned,
otherwise \href{https://fokus.rpkg.dev/articles/raw_qstnr_schema.html#supported-keys}{\code{variable_label}}. Note that \code{ballot_date} and \code{canton} either must both be
\code{NULL} or set to a valid canton name and ballot date respectively.
}
\examples{
# if `ballot_date` and `canton` are not supplied, `variable_label_common` is returned:
fokus::var_lbl("participation_federal")

# otherwise `variable_label`:
fokus::var_lbl(var_name = "participation_federal",
               ballot_date = "2018-09-23",
               canton = "aargau")
            
fokus::var_lbl(var_name = "participation_federal",
               ballot_date = "2019-10-20",
               canton = "aargau")
}
\seealso{
Other variable-related functions:
\code{\link{is_skill_question_var}()},
\code{\link{var_lvls}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()}
}
\concept{vars}
