% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_lvls}
\alias{var_lvls}
\title{Determine variable's political level(s)}
\usage{
var_lvls(var_name)
}
\arguments{
\item{var_name}{Variable name. A character scalar.}
}
\value{
A character vector of political levels, of length 0 if no specific levels could be determined.
}
\description{
Determines a variable's political level(s).
}
\details{
Note that the political levels are determined by simply parsing \code{var_name}.
}
\examples{
fokus::var_lvls("skill_question_1_cantonal_proposal_1")
}
\seealso{
Other variable-related functions:
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()}
}
\concept{vars}
