% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_prcd}
\alias{var_prcd}
\title{Determine variable's election procedure}
\usage{
var_prcd(var_names)
}
\arguments{
\item{var_names}{A character vector of FOKUS variable names.}
}
\value{
A character vector of the same length as \code{var_names}.
}
\description{
Determines the election procedure each variable corresponds to. In case no election procedure could be determined for a variable, \code{NA_character} is returned.
}
\details{
Note that the determination is performed by simply parsing \code{var_names}.
}
\examples{
fokus::qstnrs |>
  dplyr::mutate(prcd = fokus::var_prcd(variable_name)) |>
  dplyr::filter(!is.na(prcd)) |>
  dplyr::select(variable_name, prcd) |>
  unique()
}
\seealso{
Other variable-related functions:
\code{\link{add_vars_to_combos}()},
\code{\link{is_skill_question_var}()},
\code{\link{var_election_nr}()},
\code{\link{var_lbl}()},
\code{\link{var_lvls}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_title}()},
\code{\link{var_val_set}()}
}
\concept{vars}
