% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_proposal_nr}
\alias{var_proposal_nr}
\title{Determine variable's proposal number}
\usage{
var_proposal_nr(var_names)
}
\arguments{
\item{var_names}{A character vector of FOKUS variable names.}
}
\value{
An integer vector of the same length as \code{var_names}.
}
\description{
Determines the proposal number each variable corresponds to. In case no proposal number could be determined for a variable, \code{NA_character} is returned.
}
\details{
Note that the determination is performed by simply parsing \code{var_names}.
}
\seealso{
Other variable-related functions:
\code{\link{add_vars_to_combos}()},
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_lvls}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_title}()},
\code{\link{var_val_set}()}
}
\concept{vars}
