% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_skill_question_nr}
\alias{var_skill_question_nr}
\title{Determine variable's skill question number}
\usage{
var_skill_question_nr(var_names)
}
\arguments{
\item{var_names}{A character vector of FOKUS variable names.}
}
\value{
An integer vector of the same length as \code{var_names}.
}
\description{
Determines the skill question number each variable corresponds to. In case no skill question number could be determined for a variable, \code{NA_character} is
returned.
}
\details{
Note that the determination is performed by simply parsing \code{var_names}.
}
\seealso{
Other variable-related functions:
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_lvls}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_title}()}
}
\concept{vars}
