% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_title}
\alias{var_title}
\title{Get variable title}
\usage{
var_title(
  var_name,
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date)
)
}
\arguments{
\item{var_name}{Variable name. A character scalar.}

\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A character scalar.
}
\description{
Extracts a variable's title from the \link[=qstnrs]{questionnaire data}.
}
\examples{
fokus::var_title(var_name = "weight_decision",
                 ballot_date = "2018-11-25",
                 canton = "aargau")

fokus::var_title(var_name = "weight_decision",
                 ballot_date = "2019-10-20",
                 canton = "aargau")
}
\seealso{
Other variable-related functions:
\code{\link{add_vars_to_combos}()},
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_lvls}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_val_set}()}
}
\concept{vars}
