% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_val_set}
\alias{var_val_set}
\title{Get variable's value set}
\usage{
var_val_set(
  var_name,
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  lang = c(all_langs, "int")
)
}
\arguments{
\item{var_name}{Variable name. A character scalar.}

\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{lang}{Language. One of \code{"de"}, \code{"en"} or \code{"int"} for the value set's integer codes."}
}
\value{
A character scalar.
}
\description{
Extracts a variable's value set from the \link[=qstnrs]{questionnaire data} in the specified language (or its integer representation if \code{lang = "int"}).
}
\examples{
fokus::var_val_set(var_name = "favored_party",
                   ballot_date = "2018-09-23",
                   canton = "aargau")

fokus::var_val_set(var_name = "favored_party",
                   ballot_date = "2023-06-18",
                   canton = "aargau")

fokus::var_val_set(var_name = "favored_party",
                   ballot_date = "2023-06-18",
                   canton = "aargau",
                   lang = "en")

fokus::var_val_set(var_name = "favored_party",
                   ballot_date = "2023-06-18",
                   canton = "aargau",
                   lang = "int")
}
\seealso{
Other variable-related functions:
\code{\link{add_vars_to_combos}()},
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_lvls}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_title}()}
}
\concept{vars}
