% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{write_private_file}
\alias{write_private_file}
\title{Write file to private FOKUS repository}
\usage{
write_private_file(
  path,
  content,
  from_file = FALSE,
  overwrite = TRUE,
  commit_message = "auto: update file via fokus R pkg",
  branch = repo_private_default_branch,
  auth_token = pal::pkg_config_val("token_repo_private"),
  quiet = FALSE
)
}
\arguments{
\item{path}{File path relative to the repository root.}

\item{content}{File content, as a character scalar for text files, or a \link[=raw]{raw vector} for binary files. Or the path to a local file as a character scalar
if \code{from_file = TRUE}.}

\item{from_file}{Whether or not \code{content} indicates the path to a local file instead of the actual file content.}

\item{overwrite}{Whether or not to overwrite an already existing file.}

\item{commit_message}{Git commit message for file creation/update.}

\item{branch}{Git branch name to upload the file to.}

\item{auth_token}{\href{https://gitlab.com/-/user_settings/personal_access_tokens}{Personal access token of a gitlab.com account} with access to the \href{https://gitlab.com/zdaarau/private/fokus_private/}{private FOKUS repository}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
\code{path}, invisibly.
}
\description{
Uploads a file to the \href{https://gitlab.com/zdaarau/private/fokus_private/}{private FOKUS repository} via \href{https://docs.gitlab.com/ee/api/rest/}{GitLab's RESTful API (v4)}.
}
\seealso{
Other data export functions:
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_qr_codes}()},
\code{\link{export_qstnr}()}

Other private FOKUS repository functions:
\code{\link{print_private_repo_structure}()},
\code{\link{read_private_file}()},
\code{\link{url_repo_private}()}
}
\concept{data_export}
\concept{private}
