% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\docType{data}
\name{ballot_metadata}
\alias{ballot_metadata}
\title{FOKUS ballot-date-canton metadata}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
ballot_metadata
}
\description{
Returns a tibble of FOKUS-survey-related ballot-date-canton metadata, valid up until 2020-10-18.
}
\seealso{
\link{ballot_dates} \link{cantons}
}
\keyword{datasets}
