% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{ballot_title}
\alias{ballot_title}
\title{Get ballot title}
\usage{
ballot_title(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  lang = all_langs
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A character scalar.
}
\description{
Returns the ballot title consisting of the \link[=ballot_types]{ballot type} and the ballot date in German prose.
}
\examples{
fokus::ballot_title(ballot_date = "2019-10-20",
                    canton = "aargau")
}
\seealso{
Other predicate functions:
\code{\link{political_issues}()},
\code{\link{postal_dispatch_way}()},
\code{\link{response_options}()}
}
\concept{predicate_other}
