% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{ballot_type}
\alias{ballot_type}
\title{Determine ballot type}
\usage{
ballot_type(canton = cantons, ballot_date = as.character(ballot_dates))
}
\arguments{
\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"} )`
}}

\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"} )`
}}
}
\value{
The ballot type as a character scalar. One of
\itemize{
\item \code{"both_referendum_and_election"}
\item \code{"referendum"}
\item \code{"election"}
}
}
\description{
Determine ballot type
}
\seealso{
Fundamental predicates
\code{\link{is_election}()},
\code{\link{is_referendum}()}
}
\concept{fundamental}
