% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{ballot_type}
\alias{ballot_type}
\title{Determine ballot type}
\usage{
ballot_type(ballot_date = ballot_dates, canton = cantons)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
The ballot type as a character scalar. One of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
\item \code{"both_referendum_and_election"}
}
}
\description{
Determines the type of the ballot for the specified canton at the specified date.
}
\examples{
fokus::ballot_type(ballot_date = "2018-09-23",
                   canton = "aargau")
}
\seealso{
Other predicate functions
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{n_election_candidates}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{n_skill_questions}()},
\code{\link{skill_question_answer_nr}()}
}
\concept{predicate}
