% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{ballot_types}
\alias{ballot_types}
\title{Get ballot types}
\usage{
ballot_types(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A character vector.
}
\description{
Determines the \link[=all_ballot_types]{types of the ballot} covered by the FOKUS survey for the specified canton on the specified political level(s) at the
specified ballot date.
}
\examples{
fokus::ballot_types(ballot_date = "2018-09-23",
                    canton = "aargau")

# in case of no ballot types at lvl, an empty character vector is returned
fokus::ballot_types(ballot_date = "2020-10-18",
                    lvls = "federal",
                    canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()}
}
\concept{predicate_fundamental}
