% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{ballot_types}
\alias{ballot_types}
\title{Determine ballot types}
\usage{
ballot_types(ballot_date = ballot_dates, canton = cantons(ballot_date))
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}
}
\value{
A character vector of ballot types. One or more of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}
}
\description{
Determines the types of the ballot for the specified canton at the specified date.
}
\examples{
fokus::ballot_types(ballot_date = "2018-09-23",
                    canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{cantons}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()}
}
\concept{predicate_fundamental}
