% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{cantons}
\alias{cantons}
\title{List cantons that are covered in \emph{any} FOKUS survey}
\format{
A character vector.
}
\usage{
cantons
}
\description{
A vector of cantons that were part of at least one FOKUS survey up until 2021-11-28.
}
\seealso{
\code{\link{ballot_dates}} \code{\link{response_option_types}} \code{\link{q_item_keys}} \code{\link{proposal_types}}
}
\keyword{datasets}
