% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{cantons}
\alias{cantons}
\title{Get cantons covered by FOKUS survey}
\usage{
cantons(ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg))
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}
}
\value{
A character vector.
}
\description{
Determines the cantons covered by the FOKUS survey at the specified ballot date.
}
\examples{
fokus::cantons(ballot_date = "2018-09-23")

# determine all cantons covered per ballot date
library(magrittr)

fokus::all_ballot_dates \%>\%
  magrittr::set_names(., .) |>
  purrr::map_chr(fokus::cantons)
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{combos_ballot_types}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()}
}
\concept{predicate_fundamental}
