% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{cantons}
\alias{cantons}
\title{Cantons covered by FOKUS survey}
\usage{
cantons(ballot_date = all_ballot_dates)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}
}
\value{
A character vector.
}
\description{
Determines the cantons covered by the FOKUS survey at the specified ballot date.
}
\examples{
fokus::cantons(ballot_date = "2018-09-23")

# determine the cantons covered in *any* FOKUS survey
fokus::all_ballot_dates |>
  purrr::map_chr(fokus::cantons) |>
  unique()
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{has_ballot_types}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()}
}
\concept{predicate_fundamental}
