% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_elections}
\alias{combos_elections}
\title{Get election combinations}
\usage{
combos_elections(
  ballot_dates = all_ballot_dates,
  lvls = all_lvls,
  cantons = all_cantons,
  prcds = all_prcds,
  incl_prcd = TRUE,
  incl_nr = incl_prcd
)
}
\arguments{
\item{ballot_dates}{FOKUS-covered ballot date(s). One or more of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{cantons}{FOKUS-covered canton name(s). One or more of
\itemize{
\item \code{"aargau"}
}}

\item{prcds}{Election procedure(s). One or more of \code{"proportional"} and \code{"majoritarian"}.}

\item{incl_prcd}{Whether or not to include election procedures in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations.}

\item{incl_nr}{Whether or not to include election numbers in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations. Only
relevant if \code{incl_prcd = TRUE}.}
}
\value{
A list with an element per political-level and optionally -election-procedure and -election-number combination.
}
\description{
Returns a list with a metadata item per ballot date, political level and optionally election procedure and number.
}
\examples{
fokus::combos_elections(ballot_dates = "2019-10-20",
                        cantons = "aargau")

# without election numbers
fokus::combos_elections(ballot_dates = "2019-10-20",
                        cantons = "aargau",
                        incl_prcd = TRUE,
                        incl_nr = FALSE)

# without election procedures and numbers
fokus::combos_elections(ballot_dates = "2019-10-20",
                        cantons = "aargau",
                        incl_prcd = FALSE,
                        incl_nr = FALSE)
}
\seealso{
Other combination functions:
\code{\link{add_vars_to_combos}()},
\code{\link{combos_ballot_types}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()}

Other election predicate functions:
\code{\link{election_candidates}()},
\code{\link{election_name}()},
\code{\link{election_names_combined}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{past_election_date}()},
\code{\link{requires_candidate_registration}()}
}
\concept{combo}
\concept{predicate_election}
