% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_proposal_arguments}
\alias{combos_proposal_arguments}
\title{Get referendum proposal combinations for which arguments have been queried}
\usage{
combos_proposal_arguments(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  incl_side = TRUE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{incl_side}{Whether or not to include argument sides (pro/contra) in the resulting list. Setting this to \code{FALSE} potentially results in fewer
combinations.}
}
\value{
A list with an element per political-level-proposal-number and optionally -side combination.
}
\description{
Returns a list containing the political levels, proposal numbers and optionally argument sides (pro/contra) for all referendum \link[=proposal_arguments]{proposal arguments} that have been queried in the post-voting survey for the specified ballot date.
}
\examples{
fokus::combos_proposal_arguments(ballot_date = "2023-06-18",
                                 canton = "aargau")
# without sides
fokus::combos_proposal_arguments(ballot_date = "2023-06-18",
                                 canton = "aargau",
                                 incl_side = FALSE)
}
\seealso{
Other proposal predicate functions:
\code{\link{argument_proposal_nrs}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{combos_proposals}()},
\code{\link{has_proposal_arguments}()},
\code{\link{has_proposal_main_motives}()},
\code{\link{is_proposal_name_plural}()},
\code{\link{main_motive_proposal_nrs}()},
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
