% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{combos_proposals}
\alias{combos_proposals}
\title{Get proposal combinations}
\usage{
combos_proposals(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  incl_nr = TRUE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{incl_nr}{Whether or not to include proposal numbers in the resulting list. Setting this to \code{FALSE} potentially results in fewer combinations.}
}
\value{
A list with an element per political-level and optionally -proposal-number combination.
}
\description{
Returns a list containing the political levels and proposal numbers for all referendum proposals that have been covered by the FOKUS survey for the specified
canton at the specified ballot date on the specified political level(s).
}
\examples{
fokus::combos_proposals(ballot_date = "2023-06-18",
                        canton = "aargau")

# without proposal numbers
fokus::combos_proposals(ballot_date = "2023-06-18",
                        canton = "aargau",
                        incl_nr = FALSE)
}
\seealso{
Other proposal predicate functions:
\code{\link{argument_proposal_nrs}()},
\code{\link{combos_proposal_arguments}()},
\code{\link{combos_proposal_main_motives}()},
\code{\link{has_proposal_arguments}()},
\code{\link{has_proposal_main_motives}()},
\code{\link{main_motive_proposal_nrs}()},
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_name}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
