% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_name}
\alias{election_name}
\title{Get election name}
\usage{
election_name(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvl = lvls(ballot_date, canton, ballot_type = "election"),
  canton = cantons(ballot_date),
  prcd = prcds(ballot_date, lvl, canton),
  election_nr = 1L,
  lang = pal::pkg_config_val("lang"),
  type = c("short", "long", "body", "body_alt")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{prcd}{Election procedure. One of \code{"proportional"} or \code{"majoritarian"}.}

\item{election_nr}{Election number. A positive integerish scalar.}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}

\item{type}{Name type. One of \code{"short"}, \code{"long"}, \code{"body"} or \code{"body_alt"}.}
}
\value{
A character scalar.
}
\description{
Returns the name of the specified election in the specified language.
}
\examples{
fokus::election_name(ballot_date = "2019-10-20",
                     lvl = "cantonal",
                     canton = "aargau",
                     prcd = "majoritarian",
                     election_nr = 1,
                     type = "body")
}
\seealso{
Other election predicate functions:
\code{\link{combos_elections}()},
\code{\link{election_candidates}()},
\code{\link{election_names_combined}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{past_election_date}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
