% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_names_combined}
\alias{election_names_combined}
\title{Get combined elections name}
\usage{
election_names_combined(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  lang = pal::pkg_config_val("lang"),
  federal_first = TRUE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}

\item{federal_first}{Whether or not to list federal elections before cantonal ones. Only has an effect if \code{"federal" \%in\% lvls}.}
}
\value{
A character scalar, empty if no elections were held.
}
\description{
Returns the combined name of all elections at the specified date on the specified level(s) for the specified canton in the specified language.
}
\examples{
fokus::election_names_combined(ballot_date = "2019-10-20",
                               lvls = "federal",
                               canton = "aargau")

fokus::election_names_combined(ballot_date = "2019-10-20",
                               canton = "aargau")

fokus::election_names_combined(ballot_date = "2019-10-20",
                               canton = "aargau",
                               federal_first = FALSE)
}
\seealso{
Other election predicate functions:
\code{\link{combos_elections}()},
\code{\link{election_candidates}()},
\code{\link{election_name}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{past_election_date}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
