% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_names_combined}
\alias{election_names_combined}
\title{Get combined elections name}
\usage{
election_names_combined(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = cantons(ballot_date),
  lang = c("de", "en"),
  federal_first = TRUE
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}

\item{federal_first}{Whether or not to list federal elections before cantonal ones. Only relevant if \code{"federal" \%in\% lvls}.}
}
\value{
A character scalar.
}
\description{
Returns the combined name of all elections at the specified date on the specified level for the specified canton.
}
\examples{
fokus::election_names_combined(ballot_date = "2019-10-20",
                               lvls = "federal",
                               canton = "aargau")

fokus::election_names_combined(ballot_date = "2019-10-20",
                               canton = "aargau")

fokus::election_names_combined(ballot_date = "2019-10-20",
                               canton = "aargau",
                               federal_first = FALSE)
}
\seealso{
Other election predicate functions:
\code{\link{election_candidate_string}()},
\code{\link{election_candidates}()},
\code{\link{election_name}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
