% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{election_tickets}
\alias{election_tickets}
\title{Get proportional election's tickets}
\usage{
election_tickets(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "election"),
  canton = cantons(ballot_date),
  election_nr = 1L
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{Election number. A positive integerish scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns the name, number and party of all tickets for the specified proportional election.
}
\examples{
fokus::election_tickets(ballot_date = "2019-10-20",
                        lvl = "federal",
                        canton = "aargau")
}
\seealso{
Other election predicate functions:
\code{\link{combos_elections}()},
\code{\link{election_candidates}()},
\code{\link{election_names_combined}()},
\code{\link{election_name}()},
\code{\link{election_parties}()},
\code{\link{n_election_candidates}()},
\code{\link{n_election_seats}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
