% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{elections}
\alias{elections}
\title{Election data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
elections
}
\description{
A tibble containing basic election data of all FOKUS questionnaires.
}
\examples{
fokus::elections
}
\seealso{
Other datasets:
\code{\link{ballots}},
\code{\link{proposals}},
\code{\link{qstnrs}}
}
\concept{qstnr_data}
\keyword{datasets}
