% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{export_print_recipients}
\alias{export_print_recipients}
\title{Export print recipients data}
\usage{
export_print_recipients(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  upload_to_g_drive = TRUE,
  g_drive_folder =
    glue::glue("fokus/{canton}/Umfragen/Dateien für Umfrageinstitut/Print-Zielgruppen/"),
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{upload_to_g_drive}{Whether or not to upload the generated files to the Google Drive folder \code{g_drive_folder}.}

\item{g_drive_folder}{Google Drive folder to upload the generated files to. Ignored if \code{upload_to_g_drive = FALSE}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}

\item{verbose}{Whether or not to print detailed status output from \link[=upload_to_g_drive]{Google Drive file upload}.}
}
\value{
\code{NULL} if no export for the specified ballot date is possible, otherwise a \link[tibble:tbl_df-class]{tibble} of the exported data, invisibly.
}
\description{
Exports a CSV dataset containing the two columns \code{id} and \code{receives_print} to the \link[=print_private_repo_structure]{private FOKUS repository}.
}
\seealso{
Other data export functions:
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_qr_codes}()},
\code{\link{export_qstnr}()},
\code{\link{write_private_file}()}
}
\concept{data_export}
