% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{export_qr_codes}
\alias{export_qr_codes}
\title{Export QR codes with personalized survey URL}
\usage{
export_qr_codes(
  ballot_date = all_ballot_dates,
  canton = cantons(ballot_date),
  upload_to_g_drive = TRUE,
  g_drive_folder = "fokus/aargau/Umfragen/Dateien für Umfrageinstitut/QR-Codes/"
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{upload_to_g_drive}{Whether or not to upload the generated files to the Google Drive folder \code{g_drive_folder}.}

\item{g_drive_folder}{Google Drive folder to deploy the generated files to. Ignored if \code{upload_to_g_drive = FALSE}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing metadata about the contents of the created ZIP archive, invisibly.
}
\description{
Exports a ZIP file that contains a \href{https://en.wikipedia.org/wiki/QR_code}{QR code} in SVG and in EPS format for each survey participant storing the
personalized survey URL to the \link[=print_fokus_private_structure]{private FOKUS directory}.
}
