% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{fct_relabel_election_parties}
\alias{fct_relabel_election_parties}
\title{Relabel factor of election parties}
\usage{
fct_relabel_election_parties(
  fct,
  from_type = c("qstnr", "run", "short"),
  to_type = c("short", "run", "qstnr"),
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvl = lvls(ballot_date, canton, ballot_type = "election", prcds = "proportional"),
  canton = cantons(ballot_date),
  election_nr = 1L,
  past = FALSE
)
}
\arguments{
\item{fct}{Factor to be relabelled.}

\item{from_type}{Name type to convert from. One of \code{"qstnr"}, \code{"run"} or \code{"short"}.}

\item{to_type}{Name type to convert to. One of \code{"short"}, \code{"run"} or \code{"qstnr"}.}

\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{Election number. A positive integerish scalar.}

\item{past}{Whether to process the current or the predecessor election's parties.}
}
\value{
A factor.
}
\description{
Relabels a \link{factor} that contains election parties from/to the different FOKUS election party name types.
}
\examples{
d <- fokus::read_survey_data(ballot_date = "2024-10-20",
                             canton = "aargau",
                             lang = "de")
# before
levels(d$voting_decision_cantonal_proportional_election_1_party)

fct_new <- fct_relabel_election_parties(
  fct = d$voting_decision_cantonal_proportional_election_1_party,
  to_type = "short",
  ballot_date = "2024-10-20",
  lvl = "cantonal",
  canton = "aargau",
  election_nr = 1L,
  past = FALSE
)

# after
levels(d$voting_decision_cantonal_proportional_election_1_party)
}
\seealso{
\code{\link[=election_parties]{election_parties()}}
}
\concept{fcts}
