% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-package.R
\docType{package}
\name{fokus-package}
\alias{fokus}
\alias{fokus-package}
\title{fokus: Provides API-like functionality around the FOKUS Post-voting
    Surveys}
\description{
fokus allows to (pre-)process and clean the raw data, analyze
    and visualize the cleaned data, create the questionnaires and provides
    other utility functions around the FOKUS post-voting surveys by the
    Centre for Democracy Studies Aarau (ZDA) at the University of Zurich,
    Switzerland.
}
\section{Package options}{


The following package-specific R options are used to control the package's functionality:\tabular{lll}{
   \strong{name} \tab \strong{description} \tab \strong{set automatically by \code{init()}} \cr
   \code{fokus.path_wd} \tab the path to the working directory (the local instance of the \href{https://gitlab.com/c2d-zda/fokus_private}{\code{fokus_private} repository}); only set automatically for user=salim \tab ✅ \cr
   \code{fokus.global_cache_lifespan} \tab the default cache lifespan for all functions taking a \code{cache_lifespan} argument; defaults to 30 days \tab ✅ \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/zdaarau/rpkgs/fokus}
  \item Report bugs at \url{https://gitlab.com/zdaarau/rpkgs/fokus/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
