% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-package.R
\docType{package}
\name{fokus-package}
\alias{fokus}
\alias{fokus-package}
\title{fokus: Provides an API around the FOKUS Post-voting Surveys}
\description{
fokus allows to (pre-)process and clean the raw data, analyze and visualize the cleaned data, create the questionnaires and provides other utility functions around the FOKUS post-voting surveys by the Centre for Democracy Studies Aarau (ZDA) at the University of Zurich, Switzerland.
}
\section{Package configuration}{


Some of fokus's functionality is controlled via package-specific global configuration which can either be set via \href{https://rdrr.io/r/base/options.html}{R options} or \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variables} (the former take precedence). This configuration includes:\tabular{llll}{
   \strong{Description} \tab \strong{R option} \tab \strong{Environment variable} \tab \strong{Default value} \cr
   Maximal timespan to preserve the package's \href{https://pkgpins.rpkg.dev/}{pkgpins} cache. Cache entries older than this will be deleted upon package loading. \tab \code{fokus.global_max_cache_age} \tab \code{FOKUS_GLOBAL_MAX_CACHE_AGE} \tab \code{"1 day"} \cr
   Path to the working directory (the local instance of the \href{https://gitlab.com/zdaarau/private/fokus_private}{\code{fokus_private} repository}). Defaults to the current working directory. \tab \code{fokus.path_repo_private} \tab \code{FOKUS_PATH_REPO_PRIVATE} \tab  \cr
   Local filesystem path to copy the built fokus qstnr to (e.g. a Git repository that in turn deploys to a static website). \tab \code{fokus.qstnr.local_deploy_path} \tab \code{FOKUS_QSTNR_LOCAL_DEPLOY_PATH} \tab  \cr
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://fokus.rpkg.dev}
  \item \url{https://gitlab.com/zdaarau/rpkgs/fokus}
  \item Report bugs at \url{https://gitlab.com/zdaarau/rpkgs/fokus/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

Other contributors:
\itemize{
  \item  Centre for Democracy Studies Aarau (ZDA) \email{info@zdaarau.ch} [copyright holder]
}

}
\keyword{internal}
