% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{g_file_mod_time}
\alias{g_file_mod_time}
\title{Get Google Drive file modification timestamp}
\usage{
g_file_mod_time(
  g_id,
  path_gcp_service_account_key = Sys.getenv("PATH_GCP_KEY_ZDA")
)
}
\arguments{
\item{g_id}{Google Drive file ID. A character scalar.}

\item{path_gcp_service_account_key}{Path to the GCP Service Account Key JSON file. See \code{\link[=auth_g_drive_gcp]{auth_g_drive_gcp()}} for details.}
}
\value{
A \link[base:DateTimeClasses]{datetime}.
}
\description{
Get Google Drive file modification timestamp
}
\seealso{
Other Google Apps functions:
\code{\link{auth_g_drive_gcp}()},
\code{\link{auth_g_sheets_gcp}()},
\code{\link{backup_g_file}()},
\code{\link{backup_g_sheet}()},
\code{\link{upload_to_g_drive}()}
}
\concept{g_apps}
