% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{gen_q_md}
\alias{gen_q_md}
\title{Generate Markdown questionnaire}
\usage{
gen_q_md(q_tibble, incl_title = FALSE)
}
\arguments{
\item{incl_title}{Whether or not to generate an \verb{<h1>} questionnaire title at the beginning of the document. If the result is intended to be fed to Pandoc,
it's recommended to set this to \code{FALSE} and provide the title via \href{https://pandoc.org/MANUAL.html#option--metadata}{Pandoc's \code{--metadata} option} instead.}
}
\value{
A character vector.
}
\description{
Generate Markdown questionnaire
}
\seealso{
Other questionnaire generation functions
\code{\link{expand_q_tibble}()},
\code{\link{export_print_recipients}()},
\code{\link{export_q}()},
\code{\link{gen_q_tibble}()},
\code{\link{qx}}
}
\concept{q_gen}
\keyword{internal}
