% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{gen_q_tibble}
\alias{gen_q_tibble}
\title{Generate questionnaire tibble}
\usage{
gen_q_tibble(
  ballot_date = all_ballot_dates,
  canton = cantons(ballot_date),
  verbose = FALSE
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{verbose}{Whether or not to print detailed progress information during questionnaire generation. Note that it will take considerably more time when this
is set to \code{TRUE}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Generate questionnaire tibble
}
\seealso{
Other questionnaire generation functions:
\code{\link{expand_q_tibble}()},
\code{\link{gen_q_md}()}
}
\concept{q_gen}
\keyword{internal}
