% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_election}
\alias{has_election}
\title{Determine whether ballot type includes an election}
\usage{
has_election(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = all_cantons,
  prcds = all_prcds
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{prcds}{Election procedure(s). One or more of
\itemize{
\item \code{"proportional"}
\item \code{"majoritarian"}
}}
}
\value{
A logical scalar.
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) covered an election of
the specified election procedure(s).
}
\details{
Note that if more than one \code{lvls}/\code{prcds} are provided (the default), they are interpreted according to the rules of \href{https://en.wikipedia.org/wiki/Logical_disjunction}{logical disjunction}, i.e. \code{TRUE} is returned if \emph{any} \code{lvls}/\code{prcds} are included in the specified ballot.

For \href{https://en.wikipedia.org/wiki/Logical_conjunction}{logical conjunction}, use e.g. \code{has_election(lvls = "cantonal") && has_election(lvls = "federal")}.
}
\examples{
fokus::has_election(ballot_date = "2018-09-23",
                    canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()}
}
\concept{predicate_fundamental}
