% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_lvl}
\alias{has_lvl}
\title{Determine whether ballot includes a political level}
\usage{
has_lvl(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  ballot_types = all_ballot_types,
  prcds = all_prcds
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level to test for. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"} or \code{ballot_types} includes \code{"election"}.}

\item{ballot_types}{Ballot type(s). One or more of \code{"referendum"} and \code{"election"}.
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}}

\item{prcds}{Election procedure(s). One or more of \code{"proportional"} and \code{"majoritarian"}.

Only relevant if \code{ballot_types} includes \code{"election"}.}
}
\value{
A logical vector of the same length as (\code{prcds} ×) \code{ballot_types}, named after \verb{lvl(.prcds).ballot_types}.
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date covered the specified political level.
}
\examples{
fokus::has_lvl(ballot_date = "2018-09-23",
               lvl = "federal",
               canton = "aargau")

fokus::has_lvl(ballot_date = "2018-09-23",
               lvl = "federal",
               canton = "aargau",
               ballot_types = "election")

fokus::has_lvl(ballot_date = "2019-10-20",
               lvl = "federal",
               canton = "aargau",
               ballot_types = "election")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()}
}
\concept{predicate_fundamental}
