% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_lvl}
\alias{has_lvl}
\title{Determine whether ballot includes a political level}
\usage{
has_lvl(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  ballot_types = all_ballot_types
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level to test for. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{ballot_types}{Ballot type(s). One or more of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}}
}
\value{
A logical scalar.
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date covered the specified political level.
}
\examples{
fokus::has_lvl(ballot_date = "2018-09-23",
               lvl = "federal",
               canton = "aargau")

fokus::has_lvl(ballot_date = "2018-09-23",
               lvl = "federal",
               canton = "aargau",
               ballot_types = "election")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_ballot_types}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()}
}
\concept{predicate_fundamental}
