% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_proposal_nrs}
\alias{has_proposal_nrs}
\title{Determine whether ballot includes referendum proposals}
\usage{
has_proposal_nrs(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  proposal_nrs = NULL
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvls} includes \code{"cantonal"}.}

\item{proposal_nrs}{Proposals number(s) to test for. An integerish vector or \code{NULL}. If \code{NULL}, falls back to \code{\link[=n_proposals]{1:n_proposals()}} present
matching the specified parameters.}
}
\value{
Either
\itemize{
\item \strong{an unnamed logical scalar} if \code{proposal_nrs} is \code{NULL} and there was no FOKUS-covered referendum at \code{ballot_date} on any \code{lvls}, or
\item \strong{a logical vector} of the same length as \code{lvls} × \code{proposal_nrs}, \strong{named} after \code{lvls.proposal_nrs}.
}
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political levels covered the specified
referendum proposal numbers.
}
\examples{
fokus::has_proposal_nrs(ballot_date = "2018-09-23",
                        canton = "aargau")

fokus::has_proposal_nrs(ballot_date = "2018-09-23",
                        lvls = "federal",
                        canton = "aargau")
                        
fokus::has_proposal_nrs(ballot_date = "2018-09-23",
                        canton = "aargau",
                        proposal_nrs = 1:5)

fokus::has_proposal_nrs(ballot_date = "2019-10-20",
                        canton = "aargau")

fokus::has_proposal_nrs(ballot_date = "2021-11-28",
                        lvls = "cantonal",
                        canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()}
}
\concept{predicate_fundamental}
