% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_referendum}
\alias{has_referendum}
\title{Determine whether ballot type includes a referendum}
\usage{
has_referendum(
  ballot_date = all_ballot_dates,
  lvls = all_lvls,
  canton = all_cantons
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}
}
\value{
A logical scalar.
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) covered a referendum.
}
\details{
Note that if more than one \code{lvls} are provided (the default), they are interpreted according to the rules of \href{https://en.wikipedia.org/wiki/Logical_disjunction}{logical disjunction}, i.e. \code{TRUE} is returned if \emph{any} \code{lvls} are included in the specified ballot.

For \href{https://en.wikipedia.org/wiki/Logical_conjunction}{logical conjunction}, use \code{has_referendum(lvls = "cantonal") && has_referendum(lvls = "federal")}.
}
\examples{
fokus::has_referendum(ballot_date = "2018-09-23",
                      lvls = "federal",
                      canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()}
}
\concept{predicate_fundamental}
