% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{has_referendum}
\alias{has_referendum}
\title{Determine whether ballot includes a referendum}
\usage{
has_referendum(
  ballot_date = pal::pkg_config_val("ballot_date"),
  lvls = all_lvls,
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvls} includes \code{"cantonal"}.}
}
\value{
A logical vector of the same length as and named after \code{lvls}.
}
\description{
Determines whether or not the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) covered a referendum.
}
\examples{
fokus::has_referendum(ballot_date = "2018-09-23",
                      canton = "aargau")

fokus::has_referendum(ballot_date = "2018-09-23",
                      lvls = "federal",
                      canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{combos_ballot_types}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()},
\code{\link{survey_channels}()}
}
\concept{predicate_fundamental}
