% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{is_skill_question}
\alias{is_skill_question}
\title{Determine if variables are skill questions}
\usage{
is_skill_question(v_names)
}
\arguments{
\item{v_names}{A character vector of FOKUS variable names.}
}
\value{
A logical scalar.
}
\description{
Determine if variables are skill questions
}
\seealso{
FOKUS-variable-related functions
\code{\link{shorten_v_names}()},
\code{\link{v_val_lbls}()}
}
\concept{variable}
