% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{lvls}
\alias{lvls}
\title{Get ballot's political levels}
\usage{
lvls(
  ballot_date = pal::pkg_config_val("ballot_date"),
  canton = cantons(ballot_date),
  ballot_type = ballot_types(ballot_date = ballot_date, canton = canton)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{ballot_type}{Ballot type. One of \code{"referendum"} or \code{"election"}.}
}
\value{
A character vector.
}
\description{
Determines the \link[=all_lvls]{political levels} covered by the FOKUS survey for the specified canton of the specified ballot type at the specified ballot date.
}
\examples{
fokus::lvls(ballot_date = "2018-09-23",
            canton = "aargau")

fokus::lvls(ballot_date = "2018-09-23",
            canton = "aargau",
            ballot_type = "election")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{combos_ballot_types}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()},
\code{\link{survey_channels}()}
}
\concept{predicate_fundamental}
