% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{lvls}
\alias{lvls}
\title{Get ballot's political levels}
\usage{
lvls(
  ballot_date = all_ballot_dates,
  canton = all_cantons,
  ballot_types = all_ballot_types
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{canton}{A valid FOKUS canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{ballot_types}{Ballot type(s). One or more of
\itemize{
\item \code{"referendum"}
\item \code{"election"}
}}
}
\value{
A character vector of \link[=all_lvls]{political levels}.
}
\description{
Determines the political levels covered by the FOKUS survey for the specified canton at the specified ballot date.
}
\examples{
fokus::lvls(ballot_date = "2018-09-23",
            canton = "aargau")

fokus::lvls(ballot_date = "2018-09-23",
            canton = "aargau",
            ballot_types = "election")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{has_ballot_types}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()}
}
\concept{predicate_fundamental}
