% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_election_candidates}
\alias{n_election_candidates}
\title{Get number of (officially registered) majoritarian election candidates}
\usage{
n_election_candidates(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  election_nr = 1L
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{The election number. A positive integer scalar (in almost all cases \code{1L}).}
}
\value{
An integer scalar.
}
\description{
Determines the number of (officially registered) candidates of a majoritarian election at the specified ballot date on the specified political level.
}
\examples{
fokus::n_election_candidates(ballot_date = "2019-10-20",
                             lvl = "cantonal",
                             canton = "aargau",
                             election_nr = 1)
}
\seealso{
Other predicate functions
\code{\link{ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{n_skill_questions}()},
\code{\link{skill_question_answer_nr}()}
}
\concept{predicate}
