% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_election_candidates}
\alias{n_election_candidates}
\title{Get number of (officially registered) majoritarian election candidates}
\usage{
n_election_candidates(
  q_supplemental,
  level = c("cantonal", "federal"),
  election_nr = 1L,
  canton = cantons
)
}
\arguments{
\item{q_supplemental}{Supplemental date-specific questionnaire data. A list. See \code{\link[=read_q_supplemental]{read_q_supplemental()}}.}

\item{level}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{election_nr}{The election number. An integer scalar (in almost all cases \code{1L}).}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"} )`
}}
}
\value{
An integer scalar.
}
\description{
Get number of (officially registered) majoritarian election candidates
}
\examples{
n_election_candidates(q_supplemental = read_q_supplemental("2019-10-20"),
                      level = "cantonal",
                      canton = "aargau")
}
\concept{higher}
