% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_election_seats}
\alias{n_election_seats}
\title{Get number of majoritarian election seats}
\usage{
n_election_seats(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  election_nr = 1L,
  type = c("vacant", "total")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{election_nr}{Election number. A positive integer scalar (in almost all cases \code{1L}).}

\item{type}{Seat type. One of
\itemize{
\item \code{"vacant"}
\item \code{"total"}
}}
}
\value{
An integer scalar.
}
\description{
Determines the number of election seats of the specified type for the specified majoritarian election.
}
\examples{
fokus::n_election_seats(ballot_date = "2019-10-20",
                        lvl = "cantonal",
                        canton = "aargau",
                        type = "total")
}
\seealso{
Other election predicate functions:
\code{\link{election_candidate_string}()},
\code{\link{election_candidates}()},
\code{\link{election_names_combined}()},
\code{\link{election_name}()},
\code{\link{election_parties}()},
\code{\link{election_tickets}()},
\code{\link{n_election_candidates}()},
\code{\link{requires_candidate_registration}()}
}
\concept{predicate_election}
