% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_elections}
\alias{n_elections}
\title{Get number of elections}
\usage{
n_elections(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvls = all_lvls,
  canton = cantons(ballot_date),
  prcds = all_prcds
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvls}{Political level(s). One or more of \code{"cantonal"} and \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{prcds}{Election procedure(s). One or more of \code{"proportional"} and \code{"majoritarian"}.}
}
\value{
An integer vector of length \code{lvls} × \code{prcds}, named after \code{lvls.prcds}.
}
\description{
Determines the number of elections covered by the FOKUS survey for the specified canton at the specified ballot date on the specified political level(s) and
of the specified election procedure(s).
}
\examples{
fokus::n_elections(ballot_date = "2018-09-23",
                   canton = "aargau")
                   
fokus::n_elections(ballot_date = "2018-09-23",
                   lvls = "federal",
                   canton = "aargau")
                   
fokus::n_elections(ballot_date = "2018-09-23",
                   lvls = "federal",
                   canton = "aargau",
                   prcds = "majoritarian")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election_nrs}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_proposals}()},
\code{\link{prcds}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()}
}
\concept{predicate_fundamental}
