% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_elections}
\alias{n_elections}
\title{Get number of elections}
\usage{
n_elections(
  canton = cantons,
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  prcd = c("proportional", "majoritarian")
)
}
\arguments{
\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lvl}{The political level. One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{prcd}{The election procedure. One or more of
\itemize{
\item \code{"proportional"}
\item \code{"majoritarian"}
}}
}
\value{
An integer.
}
\description{
Determines the number of elections for a canton at a specific ballot date.
}
\examples{
fokus::n_elections(canton = "aargau",
                   ballot_date = "2018-09-23")
}
\seealso{
Other fundamental predicates
\code{\link{ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_referendum}()},
\code{\link{n_proposals}()}
}
\concept{fundamental}
