% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_proposal_main_motives}
\alias{n_proposal_main_motives}
\title{Get proposal's number of main motives}
\usage{
n_proposal_main_motives(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  proposal_nr = 1L,
  type = c("yes", "no")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{Canton name. One of
\itemize{
\item \code{"aargau"} .
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integer scalar.}

\item{type}{Main motive type. One of
\itemize{
\item \code{"yes"}
\item \code{"no"}
}}
}
\value{
An integer scalar.
}
\description{
Determines the number of main motives on the specified proposal.
}
\examples{
fokus::n_proposal_main_motives(ballot_date = "2018-09-23",
                               lvl = "cantonal",
                               canton = "aargau",
                               proposal_nr = 1,
                               type = "no")
}
\seealso{
Other proposal predicate functions:
\code{\link{n_proposal_arguments}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_argument}()},
\code{\link{proposal_main_motives}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_name}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
