% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_proposals}
\alias{n_proposals}
\title{Get number of referendum proposals}
\usage{
n_proposals(
  ballot_date = ballot_dates,
  lvls = c("cantonal", "federal"),
  canton = cantons
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvls}{Political level(s). One or more of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
An integer.
}
\description{
Determines the number of referendum proposals for a canton at the specified ballot date on the specified political level(s).
}
\details{
Technically, the number of \emph{federal} proposals is independent from the canton, but this function still expects a valid \code{canton}. Nonetheless, the returned
number of \emph{federal} proposals at a specific ballot date is always the same regardless of \code{canton}.
}
\examples{
fokus::n_proposals(ballot_date = "2018-09-23",
                   canton = "aargau")
}
\seealso{
Other fundamental predicate functions
\code{\link{ballot_types}()},
\code{\link{ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{n_elections}()}
}
\concept{predicate_fundamental}
