% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_skill_questions}
\alias{n_skill_questions}
\title{Get number of skill questions}
\usage{
n_skill_questions(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  proposal_nr = NULL
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}

Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integerish scalar or \code{NULL}. If \code{NULL}, the number of non-proposal-specific skill questions is returned.}
}
\value{
An integer scalar.
}
\description{
Determines the number of skill questions at the specified ballot date on the specified political level. Note that by default (\code{proposal_nr = NULL}), the
number of non-proposal-specific skill questions is returned.
}
\examples{
fokus::n_skill_questions(ballot_date = "2018-09-23",
                         lvl = "cantonal",
                         canton = "aargau",
                         proposal_nr = 1)

# note that by default, the number of non-proposal-specific skill questions is returned
fokus::n_skill_questions(ballot_date = "2018-09-23",
                         lvl = "cantonal",
                         canton = "aargau")

fokus::n_skill_questions(ballot_date = "2019-10-20",
                         lvl = "cantonal",
                         canton = "aargau")
}
\seealso{
Other skill question predicate functions:
\code{\link{skill_question_answer_nr}()},
\code{\link{skill_question_nrs}()},
\code{\link{skill_question_proposal_nrs}()},
\code{\link{skill_question_response_options}()},
\code{\link{skill_question}()}
}
\concept{predicate_skill_question}
