% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{n_skill_questions}
\alias{n_skill_questions}
\title{Get number of skill questions}
\usage{
n_skill_questions(
  canton = cantons,
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  proposal_nr = NULL
)
}
\arguments{
\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
Only relevant if \code{lvl = "cantonal"}.
}}

\item{lvl}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{proposal_nr}{The proposal number. An integer scalar or \code{NULL}. If \code{NULL}, the skill questions are considered to be non-proposal-specific (usually the
case at elections).}
}
\value{
An integer scalar.
}
\description{
Determines the number of skill questions at a specific ballot date on a specific political level.
}
\examples{
fokus::n_skill_questions(canton = "aargau",
                         ballot_date = "2018-09-23",
                         lvl = "cantonal",
                         proposal_nr = 1)
}
\seealso{
Other predicate functions
\code{\link{ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_lvl}()},
\code{\link{has_referendum}()},
\code{\link{n_election_candidates}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{skill_question_answer_nr}()}
}
\concept{predicate}
