% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{phrase}
\alias{phrase}
\title{Phrase term}
\usage{
phrase(term, vals, lang = all_langs)
}
\arguments{
\item{term}{Term to be phrased. A character scalar.}

\item{vals}{Value(s) of \code{term} to be phrased. A character vector.}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A character scalar.
}
\description{
Transforms a term's value into a phrased string representation.
}
\examples{
phrase(term = "side",
       vals = c("pro", "pro", "contra"),
       lang = "de")
}
\seealso{
Other text phrasing functions:
\code{\link{phrase_election_candidate}()},
\code{\link{phrase_proposal_name_de}()}
}
\concept{phrase}
