% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{political_issues}
\alias{political_issues}
\title{Get political issues}
\usage{
political_issues(ballot_date = ballot_dates, lang = c("de", "en"))
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character vector.
}
\description{
Returns the political issues in the specified language.
}
\examples{
fokus::political_issues(ballot_date = "2019-10-20",
                        lang = "en")
}
\concept{predicate_other}
