% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{postal_dispatch_type_prose}
\alias{postal_dispatch_type_prose}
\title{Transform postal dispatch way into prose string}
\usage{
postal_dispatch_type_prose(
  dispatch_type = all_postal_dispatch_types,
  lang = c("de", "en")
)
}
\arguments{
\item{dispatch_type}{Postal dispatch type. One of
\itemize{
\item \code{"invitation"}
\item \code{"prepaid_reply_envelope"}
\item \code{"reminder"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=all_postal_dispatch_types]{postal dispatch type} into a full prose string representation.
}
\examples{
fokus::all_postal_dispatch_types |>
  purrr::map_dfr(~ tibble::tibble(type = .x,
                                  de = fokus::postal_dispatch_type_prose(dispatch_type = .x,
                                                                         lang = "de"),
                                  en = fokus::postal_dispatch_type_prose(dispatch_type = .x,
                                                                         lang = "en")))
}
\seealso{
Other questionnaire and survey functions and datasets:
\code{\link{elections}},
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_q_all}()},
\code{\link{export_qr_codes}()},
\code{\link{export_q}()},
\code{\link{postal_dispatch_way_prose}()},
\code{\link{proposals}},
\code{\link{qx}}
}
\concept{q_survey}
