% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{postal_dispatch_type_prose}
\alias{postal_dispatch_type_prose}
\title{Transform postal dispatch type into prose string}
\usage{
postal_dispatch_type_prose(
  dispatch_type = all_postal_dispatch_types,
  lang = c("de", "en")
)
}
\arguments{
\item{dispatch_type}{Postal dispatch type. One of
\itemize{
\item \code{"invitation"}
\item \code{"prepaid_reply_envelope"}
\item \code{"reminder"}
}}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=all_postal_dispatch_types]{postal dispatch type} into a full prose string representation.
}
\examples{
fokus::all_postal_dispatch_types |>
  purrr::map(\(x) tibble::tibble(type = x,
                                 de = fokus::postal_dispatch_type_prose(dispatch_type = x,
                                                                        lang = "de"),
                                 en = fokus::postal_dispatch_type_prose(dispatch_type = x,
                                                                        lang = "en"))) |>
  purrr::list_rbind()
}
\seealso{
Other prose text functions:
\code{\link{ballot_type_prose}()},
\code{\link{postal_dispatch_way_prose}()}
}
\concept{prose}
