% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{postal_dispatch_way_prose}
\alias{postal_dispatch_way_prose}
\title{Transform postal dispatch way into prose string}
\usage{
postal_dispatch_way_prose(
  dispatch_way = all_postal_dispatch_ways,
  lang = all_langs
)
}
\arguments{
\item{dispatch_way}{Postal dispatch way One of \code{"A"}, \code{"B"} or \code{"B bulk mailing"}.}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}
}
\value{
A character scalar.
}
\description{
Transforms a \link[=postal_dispatch_way]{postal dispatch way} into a full prose string representation.
}
\examples{
fokus::postal_dispatch_way(ballot_date = "2018-09-23",
                           canton = "aargau",
                           dispatch_type = "invitation") |>
  fokus::postal_dispatch_way_prose()
}
\seealso{
Other prose text functions:
\code{\link{argument_side_prose}()},
\code{\link{ballot_type_prose}()},
\code{\link{motive_type_prose}()},
\code{\link{postal_dispatch_type_prose}()},
\code{\link{proposal_name_declined_de}()}

Other postal dispatch functions:
\code{\link{postal_dispatch_type_prose}()},
\code{\link{postal_dispatch_way}()}
}
\concept{postal_dispatch}
\concept{prose}
