% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{prcds}
\alias{prcds}
\title{Get ballot's election procedures}
\usage{
prcds(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  lvl = lvls(ballot_date, canton, ballot_type = "election"),
  canton = cantons(ballot_date)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
}}

\item{lvl}{Political level. One of \code{"cantonal"} or \code{"federal"}.}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}
}
\value{
A character vector.
}
\description{
Determines the \link[=all_prcds]{election procedures} covered by the FOKUS survey for the specified canton on the specified political level at the specified ballot
date.
}
\examples{
fokus::prcds(ballot_date = "2019-10-20",
             lvl = "cantonal",
             canton = "aargau")
             
fokus::prcds(ballot_date = "2019-10-20",
             lvl = "federal",
             canton = "aargau")

# in case of no elections, an empty character vector is returned
fokus::prcds(ballot_date = "2020-10-18",
             lvl = "federal",
             canton = "aargau")
}
\seealso{
Other fundamental predicate functions:
\code{\link{ballot_types}()},
\code{\link{cantons}()},
\code{\link{election_nrs}()},
\code{\link{election_prcds}()},
\code{\link{has_ballot_type}()},
\code{\link{has_election}()},
\code{\link{has_election_nrs}()},
\code{\link{has_lvl}()},
\code{\link{has_proposal_nrs}()},
\code{\link{has_referendum}()},
\code{\link{lvls}()},
\code{\link{n_elections}()},
\code{\link{n_proposals}()},
\code{\link{proposal_nrs}()},
\code{\link{proposal_qstn_groups}()},
\code{\link{qstn_groups_proposal_nrs}()}
}
\concept{predicate_fundamental}
