% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{prettify_date}
\alias{prettify_date}
\title{Prettify date}
\usage{
prettify_date(date, locale = "en-US")
}
\arguments{
\item{date}{The date to be prettified. A \link[base:Dates]{date} or something coercible to.}

\item{locale}{The locale the date should be prettified for. Currently only \code{"en"}/\code{"en-US"} and \code{"de"}/\code{"de-CH"} are implemented.}
}
\value{
A character scalar.
}
\description{
Note that this might only work on (Ubuntu) Linux in the current form since locales are one bitchy hell of a PITA...
}
\examples{
lubridate::today() \%>\% prettify_date()
}
