% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{proposal_main_motives}
\alias{proposal_main_motives}
\title{Get proposal's main motives}
\usage{
proposal_main_motives(
  ballot_date = ballot_dates,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  proposal_nr = 1L,
  type = c("yes", "no")
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{proposal_nr}{Proposal number. A positive integer scalar.}

\item{type}{Main motive type. One of
\itemize{
\item \code{"yes"}
\item \code{"no"}
}}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns text and code number of all main motives on the specified proposal.
}
\examples{
fokus::proposal_main_motives(ballot_date = "2018-09-23",
                             lvl = "cantonal",
                             canton = "aargau",
                             proposal_nr = 1,
                             type = "no")
}
\seealso{
Other proposal predicate functions
\code{\link{n_proposal_arguments}()},
\code{\link{n_proposal_main_motives}()},
\code{\link{proposal_arguments}()},
\code{\link{proposal_name_gender}()},
\code{\link{proposal_name}()},
\code{\link{proposal_type}()}
}
\concept{predicate_proposal}
