% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{proposals}
\alias{proposals}
\title{Proposal data}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
proposals
}
\description{
A tibble containing basic referendum proposal data of all FOKUS questionnaires.
}
\examples{
fokus::proposals
}
\seealso{
Other questionnaire and survey functions and datasets:
\code{\link{elections}},
\code{\link{export_easyvote_municipalities}()},
\code{\link{export_print_recipients}()},
\code{\link{export_q_all}()},
\code{\link{export_qr_codes}()},
\code{\link{export_q}()},
\code{\link{qx}}
}
\concept{q_survey}
\keyword{datasets}
