% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{q}
\alias{q}
\title{Raw FOKUS questionnaire data}
\format{
A \link[xfun:strict_list]{strict list}.
}
\usage{
q
}
\description{
A structured list of the raw questionnaire data of the FOKUS surveys.
}
\seealso{
\code{\link{qx_suppl}} \code{\link{q_suppl}} \code{\link{q_suppl_lvl}} \code{\link{q_suppl_lvl_canton}}
\code{\link{q_suppl_proposal}} \code{\link{q_suppl_election}} \code{\link{q_suppl_skill_questions}}
\code{\link{q_suppl_skill_question}}
}
\keyword{datasets}
