% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{q_item_keys}
\alias{q_item_keys}
\title{Questionnaire item keys}
\format{
A \link[tibble:tbl_df-class]{tibble}.
}
\usage{
q_item_keys
}
\description{
A tibble of item keys supported in the \link[=raw_q]{raw FOKUS questionnaire data}.
}
\examples{
fokus::q_item_keys
}
\seealso{
\code{\link{all_ballot_dates}} \code{\link{all_cantons}} \code{\link{all_ballot_types}} \code{\link{all_lvls}}
\code{\link{all_prcds}} \code{\link{all_proposal_types}} \code{\link{all_response_option_types}}
}
\keyword{datasets}
