% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_md}
\alias{q_md}
\title{Generate Markdown questionnaire}
\usage{
q_md(canton = cantons, ballot_date = ballot_dates)
}
\arguments{
\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
}}
}
\value{
A character vector.
}
\description{
Generate Markdown questionnaire
}
\seealso{
\code{\link{qx_supplemental}}

Other questionnaire-related functions
\code{\link{gen_q}()},
\code{\link{pick_right}()},
\code{\link{q_tibble}()}
}
\concept{questionnaire}
