% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{q_parties}
\alias{q_parties}
\title{Get political parties}
\usage{
q_parties(ballot_date = all_ballot_dates)
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Returns a tibble of ballot-date-specific political party metadata defined in the \link[=raw_q]{raw FOKUS questionnaire data}.
}
\seealso{
Other questionnaire-internal functions:
\code{\link{q_item_val}()},
\code{\link{q_response_option_codes}()}
}
\concept{q_internal}
\keyword{internal}
