% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{q_tibble}
\alias{q_tibble}
\title{Generate questionnaire tibble}
\usage{
q_tibble(canton = cantons, ballot_date = as.character(ballot_dates))
}
\arguments{
\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"} )`
}}

\item{ballot_date}{A valid FOKUS-covered cantonal ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"} )`
}}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Generate questionnaire tibble
}
\seealso{
Questionnaire-related functions
\code{\link{gen_q}()},
\code{\link{pick_right}()},
\code{\link{q_md}()},
\code{\link{read_q_supplemental}()},
\code{\link{read_q}()}
}
\concept{questionnaire}
