% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{raw_q}
\alias{raw_q}
\title{Raw FOKUS questionnaire data}
\format{
A \link[xfun:strict_list]{strict list}.
}
\description{
A structured list of the raw questionnaire data of the FOKUS surveys.
}
\seealso{
Other raw questionnaire data functions:
\code{\link{raw_pick_right}()},
\code{\link{raw_q_suppl_arguments}()},
\code{\link{raw_q_suppl_argument}()},
\code{\link{raw_q_suppl_election_name}()},
\code{\link{raw_q_suppl_elections}()},
\code{\link{raw_q_suppl_election}()},
\code{\link{raw_q_suppl_lvl_canton}()},
\code{\link{raw_q_suppl_lvl}()},
\code{\link{raw_q_suppl_main_motives}()},
\code{\link{raw_q_suppl_mode}()},
\code{\link{raw_q_suppl_proposal_name}()},
\code{\link{raw_q_suppl_proposal}()},
\code{\link{raw_q_suppl_skill_questions}()},
\code{\link{raw_q_suppl_skill_question}()},
\code{\link{raw_q_suppl}()},
\code{\link{raw_qx_suppl}}
}
\concept{q_raw}
\keyword{internal}
