% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\docType{data}
\name{raw_qx_suppl}
\alias{raw_qx_suppl}
\title{Raw supplemental date-specific FOKUS questionnaire data}
\format{
A \link[xfun:strict_list]{strict list}.
}
\usage{
raw_qx_suppl
}
\description{
A structured list of raw supplemental date-specific questionnaire data of the FOKUS surveys.
}
\examples{
fokus:::raw_qx_suppl[["2018-09-23"]]$mode
}
\seealso{
\code{\link{raw_q}} \code{\link{raw_q_suppl}} \code{\link{raw_q_suppl_lvl}} \code{\link{raw_q_suppl_lvl_canton}}
\code{\link{raw_q_suppl_proposal}} \code{\link{raw_q_suppl_election}} \code{\link{raw_q_suppl_skill_questions}}
\code{\link{raw_q_suppl_skill_question}}
}
\keyword{internal}
