% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_online_participation_codes}
\alias{read_online_participation_codes}
\title{Read in online participation codes}
\usage{
read_online_participation_codes(
  ballot_date = pal::pkg_config_val("ballot_date"),
  canton = cantons(ballot_date),
  use_cache = TRUE,
  auth_token = pal::pkg_config_val("token_repo_private")
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{use_cache}{Whether or not to return cached results if possible. Caching is done based on file content hashing, so the file is only newly fetched if it
actually changed since the last download. If \code{FALSE}, the file is always newly fetched.}

\item{auth_token}{\href{https://gitlab.com/-/user_settings/personal_access_tokens}{Personal access token of a gitlab.com account} with access to the \href{https://gitlab.com/zdaarau/private/fokus_private/}{private FOKUS repository}.}
}
\value{
A character vector.
}
\description{
Reads in the online participation codes externally generated by the survey institute that are necessary to \link[=export_qr_codes]{generate QR codes with personalized survey URLs}.
}
\examples{
# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_online_participation_codes(ballot_date = "2018-11-25",
                                         canton = "aargau") |>
    length()
)
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_private_file}()},
\code{\link{read_survey_data}()},
\code{\link{read_voting_register_data_extra}()},
\code{\link{read_voting_register_ids}()},
\code{\link{tidy_cols}()}
}
\concept{data_import}
