% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{read_survey_data}
\alias{read_survey_data}
\title{Read in generated survey data}
\usage{
read_survey_data(
  ballot_date = pal::pkg_config_val(key = "ballot_date", pkg = this_pkg),
  canton = cantons(ballot_date),
  lang = all_langs,
  merged = FALSE,
  use_cache = TRUE,
  auth_token = pal::pkg_config_val(key = "token_repo_private", pkg = this_pkg)
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
\item \code{"2023-06-18"}
\item \code{"2024-10-20"}
}}

\item{canton}{FOKUS-covered canton name. One of
\itemize{
\item \code{"aargau"}
}}

\item{lang}{Language. One of \code{"de"} or \code{"en"}.}

\item{merged}{Whether or not to read in the \emph{merged} dataset that includes the data from \strong{all} ballot dates up until \code{ballot_date} that were covered by
FOKUS surveys in the specified \code{canton}. See section \emph{Merged data} below for details. If \code{FALSE}, only the data for \code{ballot_date} itself is returned.}

\item{use_cache}{Whether or not to return cached results if possible. Caching is done based on file content hashing, so the file is only newly fetched if it
actually changed since the last download. If \code{FALSE}, the file is always newly fetched.}

\item{auth_token}{\href{https://gitlab.com/-/user_settings/personal_access_tokens}{Personal access token of a gitlab.com account} with access to the \href{https://gitlab.com/c2d-zda/private/fokus_private/}{private FOKUS repository}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Reads in the FOKUS survey dataset generated by \code{\link[=export_survey_data]{export_survey_data()}} for the specified ballot date in the specified canton and language, optionally merged
with all prior survey datasets for that canton and language.
}
\section{Merged data}{
Merged datasets differ from single-ballot-date datasets in two ways:
\enumerate{
\item Column \strong{labels} (the \code{label} attribute) are ballot-date-independent, i.e. do not include proposal or election names, and in case of question rewordings
over time, the very \emph{latest} question wording of all FOKUS surveys in that canton is used (which might have been introduced only after \code{ballot_date}). So
be aware that the questions in the labels do not necessarily correspond to the actual questions asked in the FOKUS survey for \code{ballot_date}.
\item Variables whose \strong{factor levels} vary across ballot dates, such as those capturing skill questions, are converted to type character. Be aware that these
variables usually aren't really comparable over time.
}
}

\examples{
# GitLab PAT with access to the private FOKUS repository is required for this function to work
try(
  fokus::read_survey_data(ballot_date = "2023-06-18",
                          canton = "aargau",
                          lang = "en") |>
    nrow()
)

try(
  fokus::read_survey_data(ballot_date = "2023-06-18",
                          canton = "aargau",
                          lang = "en",
                          merged = TRUE) |>
    nrow()
)
}
\seealso{
Other data import functions:
\code{\link{read_easyvote_municipalities}()},
\code{\link{read_online_participation_codes}()},
\code{\link{read_private_file}()},
\code{\link{read_voting_register_data_extra}()},
\code{\link{read_voting_register_ids}()}
}
\concept{data_import}
