% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{read_toml}
\alias{read_toml}
\title{Read in and parse a TOML file as a strict list}
\usage{
read_toml(path)
}
\arguments{
\item{path}{The path to the TOML file. A character scalar.}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Reads in a file in \href{https://toml.io/}{Tom's Obvious Minimal Language} format and returns its content as a (nested) \link[xfun:strict_list]{strict list}.
}
\details{
The file is parsed using \code{\link[RcppTOML:parseTOML]{RcppTOML::parseTOML(escape = FALSE)}}.
}
\examples{
q_path <- fokus:::path_wd("questionnaire/questionnaire.toml")

if (fs::file_exists(q_path)) fokus::read_toml(q_path)
}
\seealso{
Questionnaire-related functions
\code{\link{gen_q}()},
\code{\link{pick_right}()},
\code{\link{q_md}()},
\code{\link{q_tibble}()},
\code{\link{read_q_supplemental}()},
\code{\link{read_q}()}
}
\concept{questionnaire}
