% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{reorder_party_fct}
\alias{reorder_party_fct}
\title{Reorder a party-referencing factor according to the usual left-right spectrum}
\usage{
reorder_party_fct(fct, min_share = 0.02, lvl_below_min = "Kleinparteien")
}
\arguments{
\item{fct}{The factor to be reordered. Its levels should reference Swiss political parties.}

\item{min_share}{The share of \code{fct}'s length below which levels are lumped together to \code{lvl_below_min}. Set to \code{0} for no lumping.}

\item{lvl_below_min}{The name of the new factor level for all the old levels that fell below \code{min_share}.}
}
\value{
A factor.
}
\description{
Reorders a factor whose levels reference Swiss political parties according to the usual left-right spectrum. Levels that make less than \code{min_share} \% of all
values will be lumped together to the new factor level \code{lvl_below_min} which will be placed in the center of the left-right spectrum.
}
\concept{fct}
