% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{restore_colnames}
\alias{restore_colnames}
\title{Restore original/unshortened column names}
\usage{
restore_colnames(x)
}
\arguments{
\item{x}{A tabular data object like a data frame or tibble.}
}
\value{
\code{x} with original/unshortened column names restored.
}
\description{
Applies the inversion of \code{\link[=shorten_colnames]{shorten_colnames()}}.
}
\seealso{
\code{\link[=shorten_v_names]{shorten_v_names()}}, \code{\link[=shorten_colnames]{shorten_colnames()}}
}
