% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{skill_question}
\alias{skill_question}
\title{Get skill question}
\usage{
skill_question(
  ballot_date = all_ballot_dates,
  lvl = all_lvls,
  canton = cantons(ballot_date),
  proposal_nr = NULL,
  skill_question_nr,
  lang = c("de", "en")
)
}
\arguments{
\item{ballot_date}{A valid FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}.}

\item{proposal_nr}{Proposal number. A positive integer scalar or \code{NULL}. If \code{NULL}, it is considered to be a non-proposal-specific skill question
(the case at elections).}

\item{skill_question_nr}{Skill question number. A positive integer scalar.}

\item{lang}{Language. One of
\itemize{
\item \code{"de"}
\item \code{"en"}
}}
}
\value{
A character scalar.
}
\description{
Returns the skill question text in the specified language.
}
\examples{
fokus::skill_question(ballot_date = "2018-09-23",
                      lvl = "cantonal",
                      canton = "aargau",
                      proposal_nr = 1,
                      skill_question_nr = 2,
                      lang = "en")
}
\seealso{
Other skill question predicate functions:
\code{\link{n_skill_questions}()},
\code{\link{skill_question_answer_nr}()},
\code{\link{skill_question_response_options}()}
}
\concept{predicate_skill_question}
