% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{skill_question_answer_nr}
\alias{skill_question_answer_nr}
\title{Get correct skill question answer number}
\usage{
skill_question_answer_nr(
  q_supplemental,
  skill_question_nr,
  lvl = c("cantonal", "federal"),
  canton = cantons,
  proposal_nr = NULL
)
}
\arguments{
\item{q_supplemental}{Supplemental date-specific questionnaire data. A list. See \code{\link{qx_supplemental}}.}

\item{skill_question_nr}{The skill question number. An integer scalar.}

\item{lvl}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"}
}}

\item{proposal_nr}{The proposal number. An integer scalar.}
}
\value{
An integer scalar.
}
\description{
Note that
\itemize{
\item \code{canton} is ignored if \code{lvl = "federal"}.
\item it is considered to be a non-proposal-specific skill question if \code{proposal_nr = NULL} (usually the case at elections).
}
}
\examples{
fokus::skill_question_answer_nr(q_supplemental = qx_supplemental[["2018-09-23"]],
                                skill_question_nr = 2,
                                lvl = "cantonal",
                                canton = "aargau",
                                proposal_nr = 1)
}
