% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{skill_question_answer_nr}
\alias{skill_question_answer_nr}
\title{Get correct skill question answer number}
\usage{
skill_question_answer_nr(
  ballot_date = all_ballot_dates,
  lvl = lvls(ballot_date, canton),
  canton = cantons(ballot_date),
  proposal_nr = NULL,
  skill_question_nr
)
}
\arguments{
\item{ballot_date}{FOKUS-covered ballot date. One of
\itemize{
\item \code{"2018-09-23"}
\item \code{"2018-11-25"}
\item \code{"2019-10-20"}
\item \code{"2020-09-27"}
\item \code{"2020-10-18"}
\item \code{"2021-11-28"}
}}

\item{lvl}{Political level \link[=lvls]{covered at} \code{ballot_date} and \code{canton}.}

\item{canton}{Canton name (lowercase) \link[=cantons]{covered at} \code{ballot_date}. Only relevant if \code{lvl = "cantonal"}.}

\item{proposal_nr}{Proposal number. A positive integer scalar or \code{NULL}. If \code{NULL}, it is considered to be a non-proposal-specific skill question
(the case at elections).}

\item{skill_question_nr}{Skill question number. A positive integer scalar.}
}
\value{
An integer scalar.
}
\description{
Returns the sequential number of the correct answer for the specified skill question.
}
\examples{
fokus::skill_question_answer_nr(ballot_date = "2018-09-23",
                                lvl = "cantonal",
                                canton = "aargau",
                                proposal_nr = 1,
                                skill_question_nr = 2)
}
\seealso{
Other skill question predicate functions:
\code{\link{n_skill_questions}()},
\code{\link{skill_question_response_options}()},
\code{\link{skill_question}()}
}
\concept{predicate_skill_question}
