% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus-GEN.R
\name{skill_question_answer_nr}
\alias{skill_question_answer_nr}
\title{Get correct skill question answer number}
\usage{
skill_question_answer_nr(
  q_supplemental,
  skill_question_nr,
  level = c("cantonal", "federal"),
  canton = cantons,
  proposal_nr = NULL
)
}
\arguments{
\item{q_supplemental}{Supplemental date-specific questionnaire data. A list. See \code{\link[=read_q_supplemental]{read_q_supplemental()}}.}

\item{skill_question_nr}{The skill question number. An integer scalar.}

\item{level}{The political level. One of
\itemize{
\item \code{"cantonal"}
\item \code{"federal"}
}}

\item{canton}{A valid FOKUS canton. One of
\itemize{
\item \code{"aargau"} )`
}}

\item{proposal_nr}{The proposal number. An integer scalar.}
}
\value{
An integer scalar.
}
\description{
Note that
\itemize{
\item \code{canton} is ignored if \code{level = "federal"}.
\item it is considered to be a non-proposal-specific skill question if \code{proposal_nr = NULL} (usually the case at elections).
}
}
