% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{upload_to_g_drive}
\alias{upload_to_g_drive}
\title{Upload files to Google Drive}
\usage{
upload_to_g_drive(filepaths, g_drive_folder = "fokus_aargau/")
}
\arguments{
\item{filepaths}{Local path(s) to the file(s) to be uploaded.}

\item{g_drive_folder}{Destination path on Google Drive where the files are to be uploaded to.}
}
\value{
\code{filepaths}, invisibly.
}
\description{
Uploads one or more files to your Google Drive.
}
\seealso{
Other Google Apps functions
\code{\link{auth_g_drive_gcp}()}
}
\concept{g_apps}
