% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{url_repo_private}
\alias{url_repo_private}
\title{Assemble private FOKUS repository URL}
\usage{
url_repo_private(..., .branch = repo_private_default_branch)
}
\arguments{
\item{...}{Optional path components added to the base URL.}

\item{.branch}{Git branch name to use in URL. A character scalar. Only relevant if \code{...} is non-empty.}
}
\value{
A character scalar.
}
\description{
Assemble private FOKUS repository URL
}
\examples{
fokus:::url_repo_private("generated")
}
\seealso{
Other private FOKUS repository functions:
\code{\link{print_private_repo_structure}()},
\code{\link{read_private_file}()},
\code{\link{write_private_file}()}
}
\concept{private}
\keyword{internal}
