% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_lvls}
\alias{var_lvls}
\title{Determine variable's political level(s)}
\usage{
var_lvls(var_names)
}
\arguments{
\item{var_names}{A character vector of FOKUS variable names.}
}
\value{
A character vector of political levels, of length 0 if no specific levels could be determined.
}
\description{
Determines political level(s) of one or more variables.
}
\details{
Note that the political levels are determined by simply parsing \code{var_names}.
}
\examples{
fokus::qstnrs |>
  dplyr::mutate(lvls = purrr::map(variable_name,
                                  fokus::var_lvls)) |>
  dplyr::filter(purrr::map_lgl(lvls,
                               \(x) length(x) > 0)) |>
  dplyr::select(variable_name, lvls) |>
  tidyr::unnest_longer(col = lvls,
                       values_to = "lvl") |>
  unique()
}
\seealso{
Other variable-related functions:
\code{\link{add_vars_to_combos}()},
\code{\link{is_skill_question_var}()},
\code{\link{var_election_nr}()},
\code{\link{var_lbl}()},
\code{\link{var_prcd}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_title}()},
\code{\link{var_val_set}()}
}
\concept{vars}
