% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fokus.gen.R
\name{var_lvls}
\alias{var_lvls}
\title{Determine variable's political level(s)}
\usage{
var_lvls(var_names)
}
\arguments{
\item{var_names}{A character vector of FOKUS variable names.}
}
\value{
A character vector of political levels, of length 0 if no specific levels could be determined.
}
\description{
Determines political level(s) of one or more variables.
}
\details{
Note that the political levels are determined by simply parsing \code{var_names}.
}
\examples{
fokus::var_lvls("skill_question_1_cantonal_proposal_1")

fokus::var_lvls(c("skill_question_1_cantonal_proposal_1",
                  "skill_question_1_federal_proposal_1"))
}
\seealso{
Other variable-related functions:
\code{\link{add_vars_to_combos}()},
\code{\link{is_skill_question_var}()},
\code{\link{var_lbl}()},
\code{\link{var_proposal_nr}()},
\code{\link{var_skill_question_nr}()},
\code{\link{var_title}()},
\code{\link{var_val_set}()}
}
\concept{vars}
